% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{add_feature}
\alias{add_feature}
\title{Add a feature to a ggplot2 instance}
\usage{
add_feature(g, feature_df, feature_color, group = NULL, ...)
}
\arguments{
\item{g}{The ggplot2 instance onto which the feature will be added}

\item{feature_df}{The data frame containing the points to add to the feature}

\item{feature_color}{A hexadecimal string with which to color the feature
once added to the plot}

\item{group}{A grouping to pass along to \code{ggplot2::aes()}. This is used
for speed in the NFL and NCAA Football plotting functions}
}
\value{
A ggplot2 instance with the feature added to it
}
\description{
Add a feature to a ggplot2 instance
}
