% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{add_line_feature}
\alias{add_line_feature}
\title{Add a line feature to a ggplot2 instance}
\usage{
add_line_feature(g, feature_df, feature_color, linetype = 1, ...)
}
\arguments{
\item{g}{The ggplot2 instance onto which the feature will be added}

\item{feature_df}{The data frame containing the points to add to the feature}

\item{feature_color}{A hexadecimal string with which to color the feature
once added to the plot}

\item{linetype}{a string containing a ggplot2 linetype}
}
\value{
A ggplot2 instance with the feature added to it
}
\description{
Add a line feature to a ggplot2 instance
}
