% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-baseball.R
\name{baseball_batters_box}
\alias{baseball_batters_box}
\title{The batter's boxes on the field. This is where a batter must stand to legally
hit the ball}
\usage{
baseball_batters_box(
  batters_box_length = 0,
  batters_box_width = 0,
  batters_box_y_adj = 0,
  batters_box_thickness = 0
)
}
\arguments{
\item{batters_box_length}{The length of the batter's box (in the y direction)
measured from the outside of the chalk lines}

\item{batters_box_width}{The width of the batter's box (in the x direction)
measured from the outside of the chalk lines}

\item{batters_box_y_adj}{The shift off of center in the y direction that the
batter's box is to be moved to properly align}

\item{batters_box_thickness}{The thickness of the chalk lines that comprise
the batter's box}
}
\value{
A data frame of the batter's box
}
\description{
The batter's boxes on the field. This is where a batter must stand to legally
hit the ball
}
\keyword{internal}
