% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball.R
\name{baseball_batters_boxes}
\alias{baseball_batters_boxes}
\title{Generate the dataframe for the points that comprise the batter's boxes and
catcher's box}
\usage{
baseball_batters_boxes(g, league)
}
\arguments{
\item{g}{A ggplot2 instance on which to add the feature}

\item{league}{The league for which to draw the surface}
}
\value{
A ggplot2 instance with the batter's boxes and catcher's box added to
it
}
\description{
Generate the dataframe for the points that comprise the batter's boxes and
catcher's box
}
