% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport-baseball.R
\name{baseball_features_set_colors}
\alias{baseball_features_set_colors}
\title{Set the colors to be used for the plot. The values provided in the arguments
are the defaults, and, where specified, are the rule-book specified values.}
\usage{
baseball_features_set_colors(
  plot_background = "#395d33",
  infield_dirt = "#9b7653",
  infield_grass = "#395d33",
  pitchers_mound = "#9b7653",
  base = "#ffffff",
  pitchers_plate = "#ffffff",
  batters_box = "#ffffff",
  catchers_box = "#ffffff",
  foul_line = "#ffffff",
  running_lane = "#ffffff"
)
}
\arguments{
\item{plot_background}{A hexadecimal string representing the color to use for
this feature}

\item{infield_dirt}{A hexadecimal string representing the color to use for
this feature}

\item{infield_grass}{A hexadecimal string representing the color to use for
this feature}

\item{pitchers_mound}{A hexadecimal string representing the color to use for
this feature}

\item{base}{A hexadecimal string representing the color to use for this
feature}

\item{pitchers_plate}{A hexadecimal string representing the color to use for
this feature}

\item{batters_box}{A hexadecimal string representing the color to use for
this feature}

\item{catchers_box}{A hexadecimal string representing the color to use for
this feature}

\item{foul_line}{A hexadecimal string representing the color to use for this
feature}

\item{running_lane}{A hexadecimal string representing the color to use for
this feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Hexadecimal values are the passed vales to this function by default, but it
is also possible to use string-named values (e.g. \code{"dodgerblue"}) when
specifying.
}
\keyword{internal}
