% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-features.R
\name{baseball_home_plate}
\alias{baseball_home_plate}
\title{Home plate. This is a pentagonal shape with its back tip located at the
origin of the coordinate system. The angled sides of home plate intersect the
baselines}
\usage{
baseball_home_plate(home_plate_edge_length = 0)
}
\arguments{
\item{home_plate_edge_length}{The length of a single edge of home plate}
}
\value{
A data frame that contains the boundary of home plate
}
\description{
Home plate. This is a pentagonal shape with its back tip located at the
origin of the coordinate system. The angled sides of home plate intersect the
baselines
}
\keyword{internal}
