% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball.R
\name{baseball_infield_grass}
\alias{baseball_infield_grass}
\title{Generate the dataframe for the points that comprise home plate}
\usage{
baseball_infield_grass(g, league)
}
\arguments{
\item{g}{A ggplot2 instance on which to add the feature}

\item{league}{The league for which to draw the surface}
}
\value{
A ggplot2 instance with the infield grass added to it
}
\description{
Generate the dataframe for the points that comprise home plate
}
