% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball.R
\name{baseball_lines}
\alias{baseball_lines}
\title{Generate the dataframe for the points that comprise the baselines and the running lane}
\usage{
baseball_lines(g, league)
}
\arguments{
\item{g}{A ggplot2 instance on which to add the feature}

\item{league}{The league for which to draw the surface}
}
\value{
A ggplot2 instance with the baselines and running lane added to it
}
\description{
Generate the dataframe for the points that comprise the baselines and the running lane
}
