% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball.R
\name{baseball_mound}
\alias{baseball_mound}
\title{Generate the dataframe for the points that comprise the pitcher's mound and
pitcher's plate}
\usage{
baseball_mound(g, league)
}
\arguments{
\item{g}{A ggplot2 instance on which to add the feature}

\item{league}{The league for which to draw the surface}
}
\value{
A ggplot2 instance with the pitcher's mound and pitcher's plate added
to it
}
\description{
Generate the dataframe for the points that comprise the pitcher's mound and
pitcher's plate
}
