% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-baseball.R
\name{baseball_pitchers_mound}
\alias{baseball_pitchers_mound}
\title{Baseball Pitcher's Mound}
\usage{
baseball_pitchers_mound(pitchers_mound_radius = 0)
}
\arguments{
\item{pitchers_mound_radius}{The radius of the pitcher's mound}
}
\value{
A data frame of the pitcher's mound's bounding coordinates
}
\description{
The pitcher's mound. This is where the pitcher's plate is located, but the
pitcher's plate is not necessarily centered on the pitcher's mound
}
\keyword{internal}
