% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-baseball.R
\name{baseball_pitchers_plate}
\alias{baseball_pitchers_plate}
\title{The pitcher's plate. This is where the pitcher must throw the ball from. It's
usually a long rectangle with its front edge as its anchor point}
\usage{
baseball_pitchers_plate(pitchers_plate_length = 0, pitchers_plate_width = 0)
}
\arguments{
\item{pitchers_plate_length}{the length (x-direction) of the pitcher's plate}

\item{pitchers_plate_width}{the width (y-direction) of the pitcher's plate}
}
\value{
A data frame of the pitcher's plate's bounding coordinates
}
\description{
The pitcher's plate. This is where the pitcher must throw the ball from. It's
usually a long rectangle with its front edge as its anchor point
}
\keyword{internal}
