% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_backboard}
\alias{basketball_backboard}
\title{The backboard is the backing onto which the basket ring (created by
\code{\link[=basketball_basket_ring]{basketball_basket_ring()}}) is affixed. This will be drawn as a
rectangle on the court as the court is drawn from an aerial view}
\usage{
basketball_backboard(backboard_width = 0, backboard_thickness = 0)
}
\arguments{
\item{backboard_width}{The width of the backboard when viewed from above.
This is the x-direction dimension of the backboard when taking the point of
view of a free throw shooter}

\item{backboard_thickness}{The thickness of the backboard when viewed from
above}
}
\value{
A data frame of the bounding box of the backboard
}
\description{
The backboard is the backing onto which the basket ring (created by
\code{\link[=basketball_basket_ring]{basketball_basket_ring()}}) is affixed. This will be drawn as a
rectangle on the court as the court is drawn from an aerial view
}
\keyword{internal}
