% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_center_circle_fill}
\alias{basketball_center_circle_fill}
\title{The center circle is broken into two parts: the
\code{\link[=basketball_center_circle_outline]{basketball_center_circle_outline()}}, and the fill (this feature), which is
the court coloring inside of the inner edge of this circle}
\usage{
basketball_center_circle_fill(center_circle_radius = 0, line_thickness = 0)
}
\arguments{
\item{center_circle_radius}{The outer radius of the center circle}

\item{line_thickness}{The thickness of the line that comprises the center
circle}
}
\value{
A data frame of the boundary of the center circle. The interior of
these coordinates correspond to the filled section
}
\description{
The center circle is broken into two parts: the
\code{\link[=basketball_center_circle_outline]{basketball_center_circle_outline()}}, and the fill (this feature), which is
the court coloring inside of the inner edge of this circle
}
\keyword{internal}
