% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_center_circle_outline}
\alias{basketball_center_circle_outline}
\title{Basketball Center Circle (Outline)}
\usage{
basketball_center_circle_outline(center_circle_radius = 0, line_thickness = 0)
}
\arguments{
\item{center_circle_radius}{The outer radius of the center circle}

\item{line_thickness}{The thickness of the line that comprises the center
circle}
}
\value{
A data frame of the boundary of the center circle
}
\description{
The center circle is broken into two parts: the outline (this feature) and
the fill, which is the court coloring inside of the inner edge of this circle
}
\keyword{internal}
