% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-features.R
\name{basketball_court_apron}
\alias{basketball_court_apron}
\title{The apron of the court is the colored boundary around the exterior of some
courts. If no such colored boundary exists, this should take the same color
as the court floor}
\usage{
basketball_court_apron(
  court_length = 0,
  court_width = 0,
  court_apron_endline = 0,
  court_apron_sideline = 0,
  court_apron_to_boundary = 0,
  line_thickness = 0
)
}
\arguments{
\item{court_length}{The length of the court}

\item{court_width}{The width of the court}

\item{court_apron_endline}{The thickness of the court's apron beyond the
endline}

\item{court_apron_sideline}{The thickness of the court's apron beyond the
sideline}

\item{court_apron_to_boundary}{The distance from the inner edge of the court
apron to the outer edge of the court's boundary line (sideline and endline
will be spaced the same)}

\item{line_thickness}{The thickness of the endline and sideline}
}
\value{
A data frame of the bounding coordinates of the court apron
}
\description{
The apron of the court is the colored boundary around the exterior of some
courts. If no such colored boundary exists, this should take the same color
as the court floor
}
\keyword{internal}
