% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_endline}
\alias{basketball_endline}
\title{Basketball Endline}
\usage{
basketball_endline(court_width = 0, line_thickness = 0)
}
\arguments{
\item{court_width}{The width of the court}

\item{line_thickness}{The thickness of the endline and sideline}
}
\value{
A data frame of the bounding coordinates of the endline
}
\description{
The endline on a basketball court, also called the baseline, is located
beyond each basket. In cases where the endline is the court apron, the
endline should still be generated and its color should be set equal to the
court apron's color (see \code{\link[=basketball_court_apron]{basketball_court_apron()}} for more information on
the court apron)
}
\keyword{internal}
