% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-basketball.R
\name{basketball_features_set_colors}
\alias{basketball_features_set_colors}
\title{Basketball Feature Colors}
\usage{
basketball_features_set_colors(
  plot_background = NULL,
  defensive_half_court = "#d2ab6f",
  offensive_half_court = "#d2ab6f",
  court_apron = "#d2ab6f",
  center_circle_outline = "#000000",
  center_circle_fill = "#d2ab6f",
  division_line = "#000000",
  endline = "#000000",
  sideline = "#000000",
  two_point_range = "#d2ab6f",
  three_point_line = "#000000",
  painted_area = "#d2ab6f",
  lane_boundary = "#000000",
  free_throw_circle_outline = "#000000",
  free_throw_circle_fill = "#d2ab6f",
  free_throw_circle_dash = "#000000",
  lane_space_mark = "#000000",
  inbounding_line = "#000000",
  substitution_line = "#000000",
  baseline_lower_defensive_box = "#000000",
  lane_lower_defensive_box = "#000000",
  team_bench_line = "#000000",
  restricted_arc = "#000000",
  backboard = "#000000",
  basket_ring = "#f55b33",
  net = "#ffffff"
)
}
\arguments{
\item{plot_background}{A hexadecimal string representing the color to use for
this feature}

\item{defensive_half_court}{A hexadecimal string representing the color to
use for this feature}

\item{offensive_half_court}{A hexadecimal string representing the color to
use for this feature}

\item{court_apron}{A hexadecimal string representing the color to use for
this feature}

\item{center_circle_outline}{A hexadecimal string representing the color to
use for this feature}

\item{center_circle_fill}{A hexadecimal string representing the color to use
for this feature}

\item{division_line}{A hexadecimal string representing the color to use for
this feature}

\item{endline}{A hexadecimal string representing the color to use for this
feature}

\item{sideline}{A hexadecimal string representing the color to use for this
feature}

\item{two_point_range}{A hexadecimal string representing the color to use for
this feature}

\item{three_point_line}{A hexadecimal string representing the color to use
for this feature}

\item{painted_area}{A hexadecimal string representing the color to use for
this feature}

\item{lane_boundary}{A hexadecimal string representing the color to use for
this feature}

\item{free_throw_circle_outline}{A hexadecimal string representing the color
to use for this feature}

\item{free_throw_circle_fill}{A hexadecimal string representing the color to
use for this feature}

\item{free_throw_circle_dash}{A hexadecimal string representing the color to
use for this feature}

\item{lane_space_mark}{A hexadecimal string representing the color to use for
this feature}

\item{inbounding_line}{A hexadecimal string representing the color to use for
this feature}

\item{substitution_line}{A hexadecimal string representing the color to use
for this feature}

\item{baseline_lower_defensive_box}{A hexadecimal string representing the
color to use for this feature}

\item{lane_lower_defensive_box}{A hexadecimal string representing the color
to use for this feature}

\item{team_bench_line}{A hexadecimal string representing the color to use for
this feature}

\item{restricted_arc}{A hexadecimal string representing the color to use for
this feature}

\item{backboard}{A hexadecimal string representing the color to use for this
feature}

\item{basket_ring}{A hexadecimal string representing the color to use for
this feature}

\item{net}{A hexadecimal string representing the color to use for this
feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Set the colors to be used for the plot. The values provided in the arguments
are the defaults, and, where specified, are the rule-book specified values.
}
\details{
Hexadecimal values are the passed vales to this function by default, but it
is also possible to use string-named values (e.g. \code{"dodgerblue"}) when
specifying.
}
\keyword{internal}
