% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_free_throw_circle}
\alias{basketball_free_throw_circle}
\title{Basketball Free Throw Circle (Outline)}
\usage{
basketball_free_throw_circle(
  overhang = 0,
  free_throw_circle_radius = 0,
  line_thickness = 0
)
}
\arguments{
\item{overhang}{The arc length of the free throw circle that hangs past the
free throw line}

\item{free_throw_circle_radius}{The radius of the free throw circle}

\item{line_thickness}{The thickness of the free throw line}
}
\value{
A data frame of the bounding coordinates of the free throw circle
}
\description{
The outline of the free throw circle. The interior filling area is created
via \code{\link[=basketball_free_throw_circle_fill]{basketball_free_throw_circle_fill()}}
}
\keyword{internal}
