% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_free_throw_circle_dash}
\alias{basketball_free_throw_circle_dash}
\title{On some courts, there are a series of dashes that comprise the bottom half of
the free throw circle (e.g. the half closer to the basket). This function
generates a single dash}
\usage{
basketball_free_throw_circle_dash(
  feature_radius = 0,
  line_thickness = 0,
  start_angle = 0,
  end_angle = 0
)
}
\arguments{
\item{feature_radius}{The radius of the free throw circle}

\item{line_thickness}{The thickness of the dash}

\item{start_angle}{The angle, in \code{radians / pi}, at which the dash
should start}

\item{end_angleThe}{angle, in \code{radians / pi}, at which the dash should
end}
}
\value{
A data frame containing the bounding coordinates of a dash on the
free throw circle
}
\description{
On some courts, there are a series of dashes that comprise the bottom half of
the free throw circle (e.g. the half closer to the basket). This function
generates a single dash
}
\keyword{internal}
