% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_free_throw_circle_fill}
\alias{basketball_free_throw_circle_fill}
\title{Basketball Free Throw Circle (Interior)}
\usage{
basketball_free_throw_circle_fill(
  free_throw_circle_radius = 0,
  line_thickness = 0
)
}
\arguments{
\item{free_throw_circle_radius}{The outer radius of the free throw circle,
measured from the center of the free throw line}

\item{line_thickness}{The thickness of the outline of the free throw circle}
}
\value{
A data frame containing the bounding coordinates of the free throw
circle's semi-circular filling
}
\description{
The filled-in section of the free throw circle. The circle is the area where
a free throw shooter stands when attempting the free throw. The outline of
this area will be created separately via
\code{\link[=basketball_free_throw_circle]{basketball_free_throw_circle()}}
}
\keyword{internal}
