% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_free_throw_lane_boundary}
\alias{basketball_free_throw_lane_boundary}
\title{The lines providing the boundary to the free throw lane. When a player is
shooting a free throw, all non-shooting players must be outside of this
boundary}
\usage{
basketball_free_throw_lane_boundary(
  lane_length = 0,
  lane_width = 0,
  line_thickness = 0
)
}
\arguments{
\item{lane_length}{The length of the free throw lane}

\item{lane_width}{The width of the free throw}

\item{line_thickness}{The thickness of the free throw lane boundary}
}
\value{
A data frame of the bounding coordinates of the free throw lane
boundary
}
\description{
NOTE: This does not include lane space markings (blocks), which will be
created via \code{\link[=basketball_lane_space_mark]{basketball_lane_space_mark()}}.
}
\keyword{internal}
