% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-features.R
\name{basketball_half_court}
\alias{basketball_half_court}
\title{Each half court spans from the inner edge of the baseline to the center of
the division line, and serves as the base layer of the court plot}
\usage{
basketball_half_court(court_length = 0, court_width = 0)
}
\arguments{
\item{court_length}{The length of the court}

\item{court_width}{The width of the court}
}
\value{
A data frame of the bounding box of half a basketball court
}
\description{
Each half court spans from the inner edge of the baseline to the center of
the division line, and serves as the base layer of the court plot
}
\keyword{internal}
