% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_inbounding_line}
\alias{basketball_inbounding_line}
\title{The inbounding line is where the ball is inbounded on the sideline when
necessary. Lines drawn on the top of the court should be drawn in a top-down
direction, and lines on the bottom of the court should be drawn in the
bottom-up direction}
\usage{
basketball_inbounding_line(
  line_thickness = 0,
  in_play_ext = 0,
  out_of_bounds_ext = 0,
  drawn_direction = ""
)
}
\arguments{
\item{line_thickness}{The thickness of the inbounding line}

\item{in_play_ext}{The extension of the inbounding line into the court}

\item{out_of_bounds_ext}{The extension of the inbounding line away from the
court}

\item{drawn_direction}{A string indicating which way, in an un-rotated plot,
the line should be drawn when looking at the plot in TV View}
}
\value{
A data frame containing the bounding box of the inbounding line
}
\description{
The inbounding line is where the ball is inbounded on the sideline when
necessary. Lines drawn on the top of the court should be drawn in a top-down
direction, and lines on the bottom of the court should be drawn in the
bottom-up direction
}
\keyword{internal}
