% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_lane_space_mark}
\alias{basketball_lane_space_mark}
\title{The lane space marks, also known as the blocks, denote where non-shooting
players stand during free throws. Players may not cross these lines before
the ball touches the rim on the shot attempt}
\usage{
basketball_lane_space_mark(feature_thickness = 0, mark_depth = 0)
}
\arguments{
\item{feature_thickness}{The thickness of the lane space mark (this is the
distance in the \code{x} direction in "TV View")}

\item{mark_depth}{The distance from the exterior edge of the free throw lane
boundary that the lane space mark extends towards the sideline}
}
\value{
A data frame containing the bounding box of the lane space mark
}
\description{
The lane space marks, also known as the blocks, denote where non-shooting
players stand during free throws. Players may not cross these lines before
the ball touches the rim on the shot attempt
}
\keyword{internal}
