% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_net}
\alias{basketball_net}
\title{To make the basket ring easier to identify, the nets will also be drawn onto
the plot. They will typically be white in color}
\usage{
basketball_net(basket_ring_inner_radius = 0)
}
\arguments{
\item{basket_ring_inner_radius}{The radius of the interior of the basket ring}
}
\value{
A data frame of the net's circular outline
}
\description{
To make the basket ring easier to identify, the nets will also be drawn onto
the plot. They will typically be white in color
}
\keyword{internal}
