% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-features.R
\name{basketball_painted_area}
\alias{basketball_painted_area}
\title{The painted area is the area contained by the free throw lane (see
\code{\link[=basketball_free_throw_lane_boundary]{basketball_free_throw_lane_boundary()}} for more information on the free
throw lane)}
\usage{
basketball_painted_area(
  lane_length = 0,
  lane_width = 0,
  paint_margin = 0,
  line_thickness = 0
)
}
\arguments{
\item{lane_length}{The length of the free throw lane}

\item{lane_width}{The width of the free throw}

\item{paint_margin}{The distance from the painted area of the lane to the
free throw lane boundary lines}

\item{free_throw_lane_boundary_line_thickness}{The thickness of the line of
the free throw lane}
}
\value{
A data frame of the bounding coordinates of the free throw lane's
painted area
}
\description{
The painted area may be a different color than the rest of the two point
range area (see \code{\link[=basketball_two_point_range]{basketball_two_point_range()}} for more information on
two-point range), but may also be the same color
}
\keyword{internal}
