% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-features.R
\name{basketball_restricted_arc}
\alias{basketball_restricted_arc}
\title{The arc located in the free-throw lane is called the restricted arc. The
interior radius should be specified for this feature.}
\usage{
basketball_restricted_arc(
  feature_radius = 0,
  line_thickness = 0,
  backboard_to_center_of_basket = 0
)
}
\arguments{
\item{feature_radius}{The interior radius of the restricted arc}

\item{line_thickness}{The thickness of the restricted arc line}

\item{backboard_to_center_of_basket}{The distance from the backboard to the
center of the basket}
}
\value{
A data frame containing the bounding coordinates of the restricted
arc
}
\description{
The arc located in the free-throw lane is called the restricted arc. The
interior radius should be specified for this feature.
}
\keyword{internal}
