% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-basketball.R
\name{basketball_sideline}
\alias{basketball_sideline}
\title{The sideline on a basketball court run the full length of the court,
typically with the team bench areas and substitution areas on their exterior.
In cases where the sideline is the court apron, the sideline should still be
generated and its color should be set equal to the court apron's color (see
\code{\link[=basketball_court_apron]{basketball_court_apron()}} for more information on the court apron)}
\usage{
basketball_sideline(court_length = 0, line_thickness = 0)
}
\arguments{
\item{court_length}{The length of the court}

\item{line_thickness}{The thickness of the endline and sideline}
}
\value{
A data frame of the bounding coordinates of the sideline
}
\description{
The sideline on a basketball court run the full length of the court,
typically with the team bench areas and substitution areas on their exterior.
In cases where the sideline is the court apron, the sideline should still be
generated and its color should be set equal to the court apron's color (see
\code{\link[=basketball_court_apron]{basketball_court_apron()}} for more information on the court apron)
}
\keyword{internal}
