% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-features.R
\name{basketball_team_bench_line}
\alias{basketball_team_bench_line}
\title{Players not in the game must stay within the team bench lines unless moving
to the substitution area (see \code{\link[=substitution_line]{substitution_line()}} class)}
\usage{
basketball_team_bench_line(
  line_thickness = 0,
  extension = 0,
  drawn_direction = ""
)
}
\arguments{
\item{line_thickness}{The thickness of the team bench line}

\item{extension}{The extension of the team bench line out of the court}

\item{drawn_direction}{A string indicating which way, in an un-rotated plot,
the line should be drawn when looking at the plot in TV View}
}
\value{
A data frame containing the bounding coordinates of the team bench
line
}
\description{
Players not in the game must stay within the team bench lines unless moving
to the substitution area (see \code{\link[=substitution_line]{substitution_line()}} class)
}
\keyword{internal}
