% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cani-functions.R
\name{cani_color_league_features}
\alias{cani_color_league_features}
\title{Check to see what features of a surface can be colored}
\usage{
cani_color_league_features(league_code, sport_name = NULL)
}
\arguments{
\item{league_code}{The case-insensitive league code to be plotted}

\item{sport_name}{The name of a sport to use in the event that the
\code{league_code} supplied has more than one sport associated with it.
Default: \code{NULL}}
}
\value{
Nothing, but a message is sent to the console
}
\description{
Check to see what features of a surface can be colored
}
\examples{
cani_color_league_features("NCAA", "basketball")
}
