% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cani-functions.R
\name{cani_plot_league}
\alias{cani_plot_league}
\title{Check to see if a league can be plotted, and alert the user as to which
functions that league will work for}
\usage{
cani_plot_league(league_code)
}
\arguments{
\item{league_code}{The case-insensitive league code to be plotted}
}
\value{
Nothing, but a message is sent to the console for the user
}
\description{
Check to see if a league can be plotted, and alert the user as to which
functions that league will work for
}
\examples{
cani_plot_league('MLB')
}
