% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football-cfl-features.R
\name{cfl_feature_dead_line}
\alias{cfl_feature_dead_line}
\title{Generate the data frame for the points that comprise the dead line}
\usage{
cfl_feature_dead_line(full_surf = TRUE, rotate = FALSE, rotation_dir = "ccw")
}
\arguments{
\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the dead lines
}
\description{
Generate the data frame for the points that comprise the dead line
}
