% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football-cfl-features.R
\name{cfl_features_set_colors}
\alias{cfl_features_set_colors}
\title{Generate the list of colors for an NFL field plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_nfl()} function (or its wrapper \code{geom_football()})}
\usage{
cfl_features_set_colors(
  grass_color = "#196f0c",
  sideline_1_color = "#ffffff",
  sideline_2_color = "#ffffff",
  dead_line_color = "#ffffff",
  goal_line_color = "#ffffff",
  yard_markings_color = "#ffffff",
  directional_arrows_color = "#ffffff",
  yard_numbers_color = "#ffffff"
)
}
\arguments{
\item{grass_color}{A hexadecimal string representing the color to use for
this feature}

\item{sideline_1_color}{A hexadecimal string representing the color to use
for this feature}

\item{sideline_2_color}{A hexadecimal string representing the color to use
for this feature}

\item{dead_line_color}{A hexadecimal string representing the color to use for
this feature}

\item{goal_line_color}{A hexadecimal string representing the color to use for
this feature}

\item{yard_markings_color}{A hexadecimal string representing the color to use
for this feature}

\item{directional_arrows_color}{A hexadecimal string representing the color
to use for this feature}

\item{yard_numbers_color}{A hexadecimal string representing the color to use
for this feature}

\item{try_marking_color}{A hexadecimal string representing the color to use
for this feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Generate the list of colors for an NFL field plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_nfl()} function (or its wrapper \code{geom_football()})
}
