% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{check_data_frame_for_plot}
\alias{check_data_frame_for_plot}
\title{Check a data frame to ensure it has the minimal features for plotting}
\usage{
check_data_frame_for_plot(df)
}
\arguments{
\item{df}{A data frame to check}
}
\value{
A boolean indicating whether the minimal features needed for plotting
(a column called \code{x} and a column called \code{y}) are present
}
\description{
Check a data frame to ensure it has the minimal features for plotting
}
