% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shapes.R
\name{create_circle}
\alias{create_circle}
\title{Create a set of x and y coordinates that form a circle (or the arc of a
circle)}
\usage{
create_circle(center = c(0, 0), npoints = 1000, d = 2, start = 0, end = 2)
}
\arguments{
\item{center}{The (x, y) coordinates of the center of the circle. Default:
(0, 0)}

\item{npoints}{The number of points with which to create the circle. This
will also be the length of the resulting data frame. Default: 1000}

\item{d}{The diameter of the circle IN THE UNITS OF THE PLOT. This default
unit will be feet. Default: 2 (unit circle)}

\item{start}{The angle (in radians, divided by pi) at which to start drawing
the circle, where zero runs along the +x axis. Default: 0}

\item{end}{The angle (in radians, divided by pi) at which to stop drawing the
circle, where zero runs along the +x axis. Default: 2}
}
\value{
A dataframe containing the points needed to draw the specified circle
}
\description{
Create a set of x and y coordinates that form a circle (or the arc of a
circle)
}
