% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shapes.R
\name{create_diamond}
\alias{create_diamond}
\title{Create a set of x and y coordinates that form a diamond This function is a
wrapper on create_rectangle() above}
\usage{
create_diamond(height, width, center = c(0, 0))
}
\arguments{
\item{height}{The vertical height of the diamond}

\item{width}{The horizontal width of the diamond}

\item{center}{The center coordinates of the diamond}
}
\value{
A dataframe containing the points needed to draw the specified
diamond
}
\description{
Create a set of x and y coordinates that form a diamond This function is a
wrapper on create_rectangle() above
}
