% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-shapes.R
\name{create_line}
\alias{create_line}
\title{Create a set of \code{x} and \code{y} coordinates that form a line}
\usage{
create_line(x_start, x_end, y_start, y_end)
}
\arguments{
\item{x_start}{The value of \code{x} at the line start}

\item{x_end}{The value of \code{x} at the line end}

\item{y_start}{The value of \code{y} at the line start}

\item{y_end}{The value of \code{y} at the line end}
}
\value{
A data frame containing the points needed to draw the specified
line
}
\description{
Create a set of \code{x} and \code{y} coordinates that form a line
}
