% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shapes.R
\name{create_rectangle}
\alias{create_rectangle}
\title{Create a set of x and y coordinates that form a rectangle}
\usage{
create_rectangle(x_min, x_max, y_min, y_max)
}
\arguments{
\item{x_min}{The minimum value of x}

\item{x_max}{The maximum value of x}

\item{y_min}{The minimum value of y}

\item{y_max}{The maximum value of y}
}
\value{
A dataframe containing the points needed to draw the specified
rectangle
}
\description{
Create a set of x and y coordinates that form a rectangle
}
