% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-shapes.R
\name{create_square}
\alias{create_square}
\title{Create Square}
\usage{
create_square(side_length, center = c(0, 0))
}
\arguments{
\item{side_length}{The length of the side of the square}

\item{center}{The center coordinates of the square}
}
\value{
A data frame containing the points needed to draw the specified
square
}
\description{
Create a set of \code{x} and \code{y} coordinates that form a square. This
function is a wrapper on \code{create_rectangle()} above
}
\keyword{internal}
