% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-shapes.R
\name{create_x_shape}
\alias{create_x_shape}
\title{Create a set of \code{x} and \code{y} coordinates that form an "X"-like shape}
\usage{
create_x_shape(bar_length, bar_width, rotation = 45)
}
\arguments{
\item{bar_length}{The length of one of the bars that forms the "X" shape,
from outer edge to outer edge}

\item{bar_width}{The width of one of the bars that forms the "X" shape, from
outer edge to outer edge}

\item{rotation}{The angle of rotation (in degrees) that the resulting
"X"-like shape should be rotated}
}
\value{
A data frame containing the points needed to draw the specified "X"
}
\description{
Create a set of \code{x} and \code{y} coordinates that form an "X"-like shape
}
\keyword{internal}
