% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-curling.R
\name{curling_back_line}
\alias{curling_back_line}
\title{Curling Back Line}
\usage{
curling_back_line(line_thickness = 0, sheet_width = 0)
}
\arguments{
\item{line_thickness}{The thickness of the back line}

\item{sheet_width}{The width of the curling sheet, from side wall to side
wall}
}
\value{
A data frame containing the bounding box of the back line
}
\description{
The back line is the line in the back of the house. Its outer edge should be
used as its anchor point
}
\keyword{internal}
