% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-curling.R
\name{curling_button}
\alias{curling_button}
\title{Curling Button}
\usage{
curling_button(feature_radius = 0)
}
\arguments{
\item{feature_radius}{The radius of the button}
}
\value{
A data frame containing the bounding coordinates of the button
}
\description{
The inner-most of the concentric circles comprising the house is called the
button. This is the intersection of the tee line (see \code{\link[=curling_tee_line]{curling_tee_line()}})
and the centre line (see \code{\link[=curling_centre_line]{curling_centre_line()}})
}
\keyword{internal}
