% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-curling.R
\name{curling_courtesy_line}
\alias{curling_courtesy_line}
\title{Curling Courtesy Line}
\usage{
curling_courtesy_line(line_thickness = 0, line_length = 0)
}
\arguments{
\item{line_thickness}{The thickness of the courtesy line}

\item{line_length}{The length of the courtesy line, from the side wall
towards the centre line}
}
\value{
A data frame containing the bounding box of the courtesy line
}
\description{
The courtesy lines are where players stand during the delivery process of
each stone when the opposing team is throwing
}
\keyword{internal}
