% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-curling.R
\name{curling_hack_foothold}
\alias{curling_hack_foothold}
\title{Curling Hack (Foothold)}
\usage{
curling_hack_foothold(foothold_depth = 0, foothold_width = 0)
}
\arguments{
\item{foothold_depth}{The depth of each foothold in the hack, from the side
nearest the house to the side nearest the back board}

\item{foothold_width}{The width of each foothold in the hack, from the side
nearest the centre line to the side nearest the nearest side wall}
}
\value{
A data frame containing the bounding box of one foothold of the hack
}
\description{
The hack exits on both sides of the curling sheet between the back board and
the back line. This is where a curler pushes off from, and it should be
centered on the centre line (see \code{\link[=curling_centre_line]{curling_centre_line()}}). This function
draws one of the footholds of the hack
}
\keyword{internal}
