% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-curling.R
\name{curling_hack_line}
\alias{curling_hack_line}
\title{Curling Hack Line}
\usage{
curling_hack_line(line_thickness = 0, hack_width = 0)
}
\arguments{
\item{line_thickness}{The thickness of the hack line}

\item{hack_width}{The width of the hack, measured from the outside of one
foothold to the outside of the other}
}
\value{
A data frame containing the bounding box of the hack line
}
\description{
The hack line connects the two footholds at each hack. It should be anchored
at the terminus of the centre line (see \code{\link[=curling_centre_line]{curling_centre_line()}} for more
information)
}
\keyword{internal}
