% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-curling.R
\name{curling_hog_line}
\alias{curling_hog_line}
\title{Curling Hog Line}
\usage{
curling_hog_line(line_thickness = 0, sheet_width = 0)
}
\arguments{
\item{line_thickness}{The thickness of the hog line}

\item{sheet_width}{The width of the curling sheet, from side wall to side
wall}
}
\value{
A data frame containing the bounding box of the hog line
}
\description{
The hog line is the line that begins the Free Guard Zone at each end of the
ice. Its inner edge (relative to the nearest house) should be used as its
anchor point
}
\keyword{internal}
