% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-curling.R
\name{curling_house_ring}
\alias{curling_house_ring}
\title{Curling House Ring}
\usage{
curling_house_ring(feature_radius = 0)
}
\arguments{
\item{feature_radius}{The radius of the ring}
}
\value{
A data frame containing the bounding coordinates of the house ring
}
\description{
The house is comprised of three concentric circles outside of the button of
varying radii. This feature is designed to be each of the house rings
excluding the button
}
\keyword{internal}
