% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-curling.R
\name{curling_sheet_apron}
\alias{curling_sheet_apron}
\title{The apron of the sheet is what separates adjacent sheets, and in this context
provides a border around the outside of the sheet}
\usage{
curling_sheet_apron(
  sheet_length = 0,
  sheet_width = 0,
  apron_behind_back = 0,
  apron_along_side = 0
)
}
\arguments{
\item{sheet_length}{The length of the sheet, from back board to back board}

\item{sheet_width}{The width of the curling sheet, from side wall to side
wall}

\item{apron_behind_back}{The extension of the apron beyond the back board}

\item{apron_along_side}{The extension of the apron running along the side
walls}
}
\value{
A data frame containing the bounding coordinates of the sheet's
apron
}
\description{
The apron of the sheet is what separates adjacent sheets, and in this context
provides a border around the outside of the sheet
}
\keyword{internal}
