% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-curling.R
\name{curling_tee_line}
\alias{curling_tee_line}
\title{Curling Tee Line}
\usage{
curling_tee_line(line_thickness = 0, sheet_width = 0)
}
\arguments{
\item{line_thickness}{The thickness of the tee line}

\item{sheet_width}{The width of the curling sheet, from side wall to side
wall}
}
\value{
A data frame containing the bounding box of the tee line
}
\description{
The tee line is the line that runs through the center of the house. Its
midpoints are connected by the centre line (see \code{\link[=curling_centre_line]{curling_centre_line()}} for
more information)
}
\keyword{internal}
