% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-functions.R
\name{distance_formula}
\alias{distance_formula}
\title{Get the Euclidean distance between two points}
\usage{
distance_formula(point_1_x, point_1_y, point_2_x = 0, point_2_y = 0)
}
\arguments{
\item{point_1}{A vector containing a point's coordinates}

\item{point_2}{A vector containing a second point's coordinates}
}
\value{
The distance between the two supplied points
}
\description{
Get the Euclidean distance between two points
}
