% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-fiba-features.R
\name{fiba_feature_court_bground}
\alias{fiba_feature_court_bground}
\title{Generate the data frame for the points that comprise a regulation FIBA
basketball court}
\usage{
fiba_feature_court_bground(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw",
  unit = "ft"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not this feature is needed
for a full-surface representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}

\item{unit}{A string indicating the unit with which to make the plot.
Default: \code{'ft'}}
}
\value{
A data frame containing the points that comprise the court
}
\description{
Generate the data frame for the points that comprise a regulation FIBA
basketball court
}
