% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-fiba-features.R
\name{fiba_feature_free_throw_lane_lines}
\alias{fiba_feature_free_throw_lane_lines}
\title{Generate the data frames for the points that comprise the free-throw lane
lines (the blocks)}
\usage{
fiba_feature_free_throw_lane_lines(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw",
  unit = "ft"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not this feature is needed
for a full-surface representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}

\item{unit}{A string indicating the unit with which to make the plot.
Default: \code{'ft'}}
}
\value{
A list of data frames containing the points that comprise the
free-throw lane lines (the blocks)
}
\description{
Generate the data frames for the points that comprise the free-throw lane
lines (the blocks)
}
