% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-fiba-features.R
\name{fiba_features_set_colors}
\alias{fiba_features_set_colors}
\title{Generate the list of colors for a FIBA court plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_fiba()} function (or its wrapper \code{geom_basketball()})}
\usage{
fiba_features_set_colors(
  court_background_color = "#d2ab6f",
  center_circle_color = "#000000",
  division_line_color = "#000000",
  endline_color = "#000000",
  sideline_color = "#000000",
  team_bench_color = "#000000",
  court_apron_color = "#d2ab6f",
  three_point_line_color = "#000000",
  two_point_range_color = "#d2ab6f",
  free_throw_lane_color = "#000000",
  painted_area_color = "#d2ab6f",
  free_throw_lane_lines_color = "#000000",
  free_throw_semi_circle_line_color = "#000000",
  free_throw_semi_circle_fill_color = "#d2ab6f",
  restricted_area_arc_color = "#000000",
  backboard_color = "#000000",
  basket_ring_color = "#000000",
  net_color = "#ffffff"
)
}
\arguments{
\item{court_background_color}{A hexadecimal string representing the color to
use for this feature}

\item{center_circle_color}{A hexadecimal string representing the color to use
for this feature}

\item{division_line_color}{A hexadecimal string representing the color to use
for this feature}

\item{endline_color}{A hexadecimal string representing the color to use for
this feature}

\item{sideline_color}{A hexadecimal string representing the color to use for
this feature}

\item{team_bench_color}{A hexadecimal string representing the color to use
for this feature}

\item{court_apron_color}{A hexadecimal string representing the color to use
for this feature}

\item{three_point_line_color}{A hexadecimal string representing the color to
use for this feature}

\item{two_point_range_color}{A hexadecimal string representing the color to
use for this feature}

\item{free_throw_lane_color}{A hexadecimal string representing the color to
use for this feature}

\item{painted_area_color}{A hexadecimal string representing the color to use
for this feature}

\item{free_throw_lane_lines_color}{A hexadecimal string representing the
color to use for this feature}

\item{free_throw_semi_circle_line_color}{A hexadecimal string representing
the color to use for this feature}

\item{free_throw_semi_circle_fill_color}{A hexadecimal string representing
the color to use for this feature}

\item{restricted_area_arc_color}{A hexadecimal string representing the color
to use for this feature}

\item{backboard_color}{A hexadecimal string representing the color to use for
this feature}

\item{basket_ring_color}{A hexadecimal string representing the color to use
for this feature}

\item{net_color}{A hexadecimal string representing the color to use for this
feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Generate the list of colors for a FIBA court plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_fiba()} function (or its wrapper \code{geom_basketball()})
}
