% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer-fifa-features.R
\name{fifa_feature_halfway_line}
\alias{fifa_feature_halfway_line}
\title{Generate the data frame for the points that comprise the halfway line}
\usage{
fifa_feature_halfway_line(
  goal_line_length,
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{goal_line_length}{The length of the goal line}

\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the halfway line
}
\description{
Generate the data frame for the points that comprise the halfway line
}
