% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimming_FINA_features.R
\name{fina_swimming_features_set_colors}
\alias{fina_swimming_features_set_colors}
\title{Generate the list of colors for a pool. The defaults can
be overwritten by supplying the names of the list elements to the
\code{geom_swimming()} function}
\usage{
fina_swimming_features_set_colors(
  deck_color = "grey",
  pool_color = "blue",
  m15_start_color = "black",
  m15_turn_color = "black",
  center_line_color = "black",
  flags_start_color = "red",
  flags_turn_color = "red",
  flags_string_color = "black",
  lane_markers_color = "black",
  blocks_color = "white",
  lane_lines_color = "white",
  lane_lines_ends_color = "red",
  lane_lines_15m_color = "red"
)
}
\arguments{
\item{deck_color}{A hexadecimal string representing the color to use for
this feature}

\item{pool_color}{A hexadecimal string representing the color to use for
this feature}

\item{m15_start_color}{A hexadecimal string representing the color to use for
this feature}

\item{m15_turn_color}{A hexadecimal string representing the color to use for
this feature}

\item{center_line_color}{A hexadecimal string representing the color to use for
this feature}

\item{flags_start_color}{A hexadecimal string representing the color to use for
this feature}

\item{flags_turn_color}{A hexadecimal string representing the color to use for
this feature}

\item{flags_string_color}{A hexadecimal string representing the color to use for
this feature}

\item{lane_markers_color}{A hexadecimal string representing the color to use for
this feature}

\item{blocks_color}{A hexadecimal string representing the color to use for
this feature}

\item{lane_lines_color}{A hexadecimal string representing the color to use for
this feature}

\item{lane_lines_ends_color}{A hexadecimal string representing the color to use for
this feature}

\item{lane_lines_15m_color}{A hexadecimal string representing the color to use for
this feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Generate the list of colors for a pool. The defaults can
be overwritten by supplying the names of the list elements to the
\code{geom_swimming()} function
}
