% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_coaching_box}
\alias{football_coaching_box}
\title{The coaching box is the area between the restricted area and team bench area.
It may or may not be distinct from either of these areas, but is typically
separated by the coaching box line (see \code{\link[=football_coaching_box_line]{football_coaching_box_line()}} for
more information on the coaching box line)}
\usage{
football_coaching_box(coaching_box_length = 0, feature_thickness = 0)
}
\arguments{
\item{coaching_box_length}{The length of the coaching box}

\item{feature_thickness}{The depth beyond the exterior edge of the restricted
area that the coaching box protrudes}
}
\value{
A data frame of the bounding box of coaching box
}
\description{
The coaching box is the area between the restricted area and team bench area.
It may or may not be distinct from either of these areas, but is typically
separated by the coaching box line (see \code{\link[=football_coaching_box_line]{football_coaching_box_line()}} for
more information on the coaching box line)
}
\keyword{internal}
