% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_coaching_box_line}
\alias{football_coaching_box_line}
\title{The coaching box line is the line that separates the team bench area from the
coaching box. This line should be a different color than the bench area and
coaching box, which may be the same color}
\usage{
football_coaching_box_line(coaching_box_line_length = 0, feature_thickness = 0)
}
\arguments{
\item{coaching_box_line_length}{The length of the line forming the coaching
box line}

\item{feature_thickness}{The thickness with which to draw the line}
}
\value{
A data frame of the bounding box of the coaching box line
}
\description{
The coaching box line is the line that separates the team bench area from the
coaching box. This line should be a different color than the bench area and
coaching box, which may be the same color
}
\keyword{internal}
