% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football.R
\name{football_directional_arrows_df_maker}
\alias{football_directional_arrows_df_maker}
\title{Generate the dataframes for the points that comprise the directional arrows. This
function is fed into the football_directional_arrows() function below and helps to
optimize performance}
\usage{
football_directional_arrows_df_maker(
  yardage,
  league,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{yardage}{The yardage for which to create the yard markings}

\item{league}{The league for which to draw the surface}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: FALSE}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: 'ccw'}
}
\value{
A dataframe (or list of dataframes) that contain the coordinates for
the directional arrows
}
\description{
Generate the dataframes for the points that comprise the directional arrows. This
function is fed into the football_directional_arrows() function below and helps to
optimize performance
}
