% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_end_line}
\alias{football_end_line}
\title{The end line is the line beyond the back of the endzone. Its interior edge is
considered out of bounds}
\usage{
football_end_line(feature_thickness = 0, field_width = 0)
}
\arguments{
\item{feature_thickness}{The thickness of the boundary lines}

\item{field_width}{The width of the field}
}
\value{
A data frame of the bounding box of the end line
}
\description{
The end line is the line beyond the back of the endzone. Its interior edge is
considered out of bounds
}
\keyword{internal}
