% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-football.R
\name{football_features_set_colors}
\alias{football_features_set_colors}
\title{Set the colors to be used for the plot. The values provided in the arguments
are the defaults, and, where specified, are the rule-book specified values.}
\usage{
football_features_set_colors(
  plot_background = NULL,
  field_apron = "#196f0c",
  offensive_half = "#196f0c",
  defensive_half = "#196f0c",
  offensive_endzone = "#196f0c",
  defensive_endzone = "#196f0c",
  end_line = "#ffffff",
  sideline = "#ffffff",
  field_border = "#196f0c",
  field_border_outline = "#ffffff",
  red_zone_border = "#196f0c",
  red_zone_border_outline = "#ffffff",
  major_yard_line = "#ffffff",
  goal_line = "#ffffff",
  minor_yard_line = "#ffffff",
  directional_arrow = "#ffffff",
  try_mark = "#ffffff",
  yardage_marker = "#ffffff",
  restricted_area = "#ffffff",
  coaching_box = "#ffffff",
  team_bench_area = "#196f0c",
  team_bench_area_outline = "#ffffff",
  coaching_box_line = "#ffcb05"
)
}
\arguments{
\item{plot_background}{A hexadecimal string representing the color to use for
this feature}

\item{field_apron}{A hexadecimal string representing the color to use for
this feature}

\item{offensive_half}{A hexadecimal string representing the color to use for
this feature}

\item{defensive_half}{A hexadecimal string representing the color to use for
this feature}

\item{offensive_endzone}{A hexadecimal string representing the color to use
for this feature}

\item{defensive_endzone}{A hexadecimal string representing the color to use
for this feature}

\item{end_line}{A hexadecimal string representing the color to use for this
feature}

\item{sideline}{A hexadecimal string representing the color to use for this
feature}

\item{field_border}{A hexadecimal string representing the color to use for
this feature}

\item{field_border_outline}{A hexadecimal string representing the color to
use for this feature}

\item{red_zone_border}{A hexadecimal string representing the color to use for
this feature}

\item{red_zone_border_outline}{A hexadecimal string representing the color to
use for this feature}

\item{major_yard_line}{A hexadecimal string representing the color to use for
this feature}

\item{goal_line}{A hexadecimal string representing the color to use for this
feature}

\item{minor_yard_line}{A hexadecimal string representing the color to use for
this feature}

\item{directional_arrow}{A hexadecimal string representing the color to use
for this feature}

\item{try_mark}{A hexadecimal string representing the color to use for this
feature}

\item{yardage_marker}{A hexadecimal string representing the color to use for
this feature}

\item{restricted_area}{A hexadecimal string representing the color to use for
this feature}

\item{coaching_box}{A hexadecimal string representing the color to use for
this feature}

\item{team_bench_area}{A hexadecimal string representing the color to use for
this feature}

\item{team_bench_area_outline}{A hexadecimal string representing the color to
use for this feature}

\item{coaching_box_line}{A hexadecimal string representing the color to use
for this feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Hexadecimal values are the passed vales to this function by default, but it
is also possible to use string-named values (e.g. \code{"dodgerblue"}) when
specifying.
}
\keyword{internal}
