% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_half_field}
\alias{football_half_field}
\title{Each half of the football field spans from the back edge of the endzone to
the center of the major yard line at midfield}
\usage{
football_half_field(field_length = 0, field_width = 0, endzone_length = 0)
}
\arguments{
\item{field_length}{The length of the field}

\item{field_width}{The width of the field}

\item{endzone_length}{The length of the endzone}
}
\value{
A data frame of the bounding box of half a football field
}
\description{
Each half of the football field spans from the back edge of the endzone to
the center of the major yard line at midfield
}
\keyword{internal}
