% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_major_yard_line}
\alias{football_major_yard_line}
\title{The major yard lines are the yard lines that span the entire width of the
field. Typically, these lines are placed every 5 yards, but the customization
is left to the user. These lines may feature a cross-hash, which runs in the
x-direction}
\usage{
football_major_yard_line(
  field_width = 0,
  feature_thickness = 0,
  dist_to_sideline = 0,
  cross_hash_length = 0,
  cross_hash_separation = 0
)
}
\arguments{
\item{field_width}{The width of the field}

\item{feature_thickness}{The thickness of each of the major yard lines}

\item{dist_to_sideline}{The distance from the end of the yard line to the
interior edge of the sideline}

\item{cross_hash_length}{The length of each cross-hash mark}

\item{cross_hash_separation}{The interior separation between the cross-hashes}
}
\value{
A data frame containing the bounding box of the major yard lines
}
\description{
The major yard lines are the yard lines that span the entire width of the
field. Typically, these lines are placed every 5 yards, but the customization
is left to the user. These lines may feature a cross-hash, which runs in the
x-direction
}
\keyword{internal}
