% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_minor_yard_line}
\alias{football_minor_yard_line}
\title{The minor yard lines are the yard lines in between all of the major yard
lines. Typically, there are four sets of minor yard lines: one near each
sideline, and two near the middle of the field}
\usage{
football_minor_yard_line(yard_line_height = 0, feature_thickness = 0)
}
\arguments{
\item{yard_line_height}{The height (in the y-direction) of each yard line}

\item{feature_thickness}{The thickness of each minor yard line}
}
\value{
A data frame of the bounding box of a minor yard line
}
\description{
The minor yard lines are the yard lines in between all of the major yard
lines. Typically, there are four sets of minor yard lines: one near each
sideline, and two near the middle of the field
}
\keyword{internal}
