% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_red_zone_border}
\alias{football_red_zone_border}
\title{The field border (see \code{\link[=football_field_border]{football_field_border()}}) may have a different color
along the red zone than it does along the rest of the field. This is not
always the case, but the feature is provided for convenience}
\usage{
football_red_zone_border(feature_thickness = 0)
}
\arguments{
\item{feature_thickness}{The thickness of the field border}
}
\value{
A data frame of the bounding box of the red zone border
}
\description{
The field border (see \code{\link[=football_field_border]{football_field_border()}}) may have a different color
along the red zone than it does along the rest of the field. This is not
always the case, but the feature is provided for convenience
}
\keyword{internal}
