% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_red_zone_border_outline}
\alias{football_red_zone_border_outline}
\title{The outline of the \code{\link[=football_red_zone_border]{football_red_zone_border()}} may be a different color than
the rest of the field border. This is not always the case, but the feature is
provided for convenience}
\usage{
football_red_zone_border_outline(feature_thickness = 0)
}
\arguments{
\item{feature_thickness}{The thickness of the field border's outline}
}
\value{
A data frame of the bounding box of the red zone border's outline
}
\description{
The outline of the \code{\link[=football_red_zone_border]{football_red_zone_border()}} may be a different color than
the rest of the field border. This is not always the case, but the feature is
provided for convenience
}
\keyword{internal}
