% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_restricted_area}
\alias{football_restricted_area}
\title{The restricted area is the area nearest the sideline's exterior edge. This
area is distinct from the coaching box (immediately behind the restricted
area) and team bench area}
\usage{
football_restricted_area(restricted_area_length = 0, feature_thickness = 0)
}
\arguments{
\item{restricted_area_length}{The length of the restricted area}

\item{feature_thickness}{The depth beyond the exterior edge of the sideline
that the restricted area protrudes}
}
\value{
A data frame of the bounding coordinates of the restricted area
}
\description{
The restricted area is the area nearest the sideline's exterior edge. This
area is distinct from the coaching box (immediately behind the restricted
area) and team bench area
}
\keyword{internal}
