% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_sideline}
\alias{football_sideline}
\title{Football Sideline}
\usage{
football_sideline(feature_thickness = 0, field_length = 0, endzone_length = 0)
}
\arguments{
\item{feature_thickness}{The thickness of the boundary lines}

\item{field_length}{The length of the field}

\item{endzone_length}{The length of the endzone}
}
\value{
A data frame of the bounding box of the sidelines
}
\description{
The sidelines are the lines that run the length of the field, stretching from
the back of one endzone to the back of the other endzone. Its interior edge
is considered out of bounds
}
\keyword{internal}
