% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_team_bench_area}
\alias{football_team_bench_area}
\title{The team bench area is the area beyond the restricted area and coaching box.
It is where the team benches, non-playing players, and team staff are to
remain during the game}
\usage{
football_team_bench_area(
  team_bench_length_field_side = 0,
  team_bench_length_back_side = 0,
  team_bench_width = 0
)
}
\arguments{
\item{team_bench_length_field_side}{The length of the side of the team bench
area closest to the field}

\item{team_bench_length_back_side}{The length of the side of the team bench
area furthest from the field}

\item{team_bench_width}{The depth beyond the outer edge of the coaching box
line that the team bench area protrudes}
}
\value{
A data frame of the bounding coordinates of the team bench area
}
\description{
The team bench area is the area beyond the restricted area and coaching box.
It is where the team benches, non-playing players, and team staff are to
remain during the game
}
\keyword{internal}
