% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-football.R
\name{football_try_mark}
\alias{football_try_mark}
\title{The try mark is the mark from which all tries start. This line is located
directly on the line \code{y = 0}. This line is not typically considered an
official yard line, which is why it is created independently}
\usage{
football_try_mark(try_mark_width = 0, feature_thickness = 0)
}
\arguments{
\item{try_mark_width}{The width (in the y-direction) of the try mark}

\item{feature_thickness}{The thickness of the try mark}
}
\value{
A data frame of the bounding box of the try mark
}
\description{
The try mark is the mark from which all tries start. This line is located
directly on the line \code{y = 0}. This line is not typically considered an
official yard line, which is why it is created independently
}
\keyword{internal}
