% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football.R
\name{football_yard_numbers}
\alias{football_yard_numbers}
\title{Add the yardline-marking numbers at 10-yard intervals}
\usage{
football_yard_numbers(g, league, rotate = FALSE, rotation_dir = "ccw")
}
\arguments{
\item{g}{A ggplot2 instance on which to add the feature}

\item{league}{The league for which to draw the surface}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: FALSE}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: 'ccw'}
}
\value{
A ggplot2 instance with the yardline-marking numbers added to it
}
\description{
Add the yardline-marking numbers at 10-yard intervals
}
