% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball.R
\name{free_throw_lane}
\alias{free_throw_lane}
\title{Generate the dataframe for the points that comprise the free-throw lane}
\usage{
free_throw_lane(
  g,
  league,
  include_amateur = FALSE,
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{g}{A ggplot2 instance on which to add the feature}

\item{league}{The league for which to draw the surface}

\item{include_amateur}{A boolean indicating whether or not to include amateur free-throw lane markings (some professional (W)NBA courts include amateur free-throw lane markings)}

\item{full_surf}{A boolean indicating whether or not this feature is needed for a full-surface representation of the surface. Default: TRUE}

\item{rotate}{A boolean indicating whether or not this feature needs to be rotated Default: FALSE}

\item{rotation_dir}{A string indicating which direction to rotate the feature. Default: 'ccw'}
}
\value{
A ggplot2 instance with the free-throw lane added to it
}
\description{
Generate the dataframe for the points that comprise the free-throw lane
}
