% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.R
\name{ft_to_yd}
\alias{ft_to_yd}
\title{Converts feet to yards}
\usage{
ft_to_yd(ft)
}
\arguments{
\item{ft}{A number of feet to be converted to yards}
}
\value{
The converted number of feet to yards
}
\description{
Converts feet to yards
}
