% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball.R
\name{geom_baseball}
\alias{geom_baseball}
\title{Generate a ggplot2 instance containing a regulation baseball field for a
specified league}
\usage{
geom_baseball(league)
}
\arguments{
\item{league}{The league for which to draw the surface}
}
\value{
A ggplot2 instance with a full-surface representation of a baseball
field
}
\description{
Generate a ggplot2 instance containing a regulation baseball field for a
specified league
}
\examples{
geom_baseball(league = "MLB")
}
