% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football-cfl-features.R
\name{geom_cfl}
\alias{geom_cfl}
\title{Create a ggplot2 instance that represents a regulation NFL field, with the
bottom left corner of the left-most endzone corresponding to  (0, 0)}
\usage{
geom_cfl(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw",
  unit = "yd",
  origin_bottom_left = TRUE,
  caption_color = "#707372",
  background_color = NULL,
  ...
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not to draw a full-surface
representation of the playing surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not the surface representation
needs to be rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the surface
representation. Default: \code{'ccw'}}

\item{unit}{A string indicating the units with which to make the plot.
Default: \code{'yd'}}

\item{origin_bottom_left}{A boolean indicating whether or not to place the
origin at the bottom-left corner of the plot (where the dead_line meets the
sideline). Default: \code{TRUE}}

\item{caption_color}{A hexadecimal string representing the color to use for
the plot's caption. Default: '#707372' (grey)}

\item{background_color}{A hexadecimal string representing the color to use
for the plot's background. Default: \code{NULL}}

\item{...}{Additional arguments to pass to the function. These should be the
colors to pass to the \code{cfl_features_set_colors()} function}
}
\value{
A ggplot2 instance that represents a regulation NFL field
}
\description{
Create a ggplot2 instance that represents a regulation NFL field, with the
bottom left corner of the left-most endzone corresponding to  (0, 0)
}
