% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-mlb-features.R
\name{geom_mlb}
\alias{geom_mlb}
\title{Create a ggplot2 instance that represents a regulation NBA court, with the
back tip of home plate corresponding to (0, 0)}
\usage{
geom_mlb(
  caption_color = "#707372",
  background_color = "#395d33",
  unit = "ft",
  ...
)
}
\arguments{
\item{caption_color}{A hexadecimal string representing the color to use for
the plot's caption. Default: '#ffffff' (white)}

\item{background_color}{A hexadecimal string representing the color to use
for the plot's background. Default: '#395d33' (green)}

\item{unit}{A string indicating the units with which to make the plot.
Default: \code{'ft'}}

\item{...}{Additional arguments to pass to the function. These should be the
colors to pass to the \code{mlb_features_set_colors()} function}
}
\value{
A ggplot2 instance that represents a regulation NBA court
}
\description{
Create a ggplot2 instance that represents a regulation NBA court, with the
back tip of home plate corresponding to (0, 0)
}
