% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey-nhl-features.R
\name{geom_nhl}
\alias{geom_nhl}
\title{Create a ggplot2 instance that represents a regulation NHL rink, with the
center of the rink corresponding to (0, 0)}
\usage{
geom_nhl(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw",
  unit = "ft",
  caption_color = "#707372",
  background_color = NULL,
  ...
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not to draw a full-surface
representation of the playing surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not the surface representation
needs to be rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the surface
representation. Default: \code{'ccw'}}

\item{unit}{A string indicating the units with which to make the plot.
Default: \code{'ft'}}

\item{caption_color}{A hexadecimal string representing the color to use for
the plot's caption. Default: '#707372' (grey)}

\item{background_color}{A hexadecimal string representing the color to use
for the plot's background. Default: \code{NULL}}

\item{...}{Additional arguments to pass to the function. These should be the
colors to pass to the \code{nhl_features_set_colors()} function}
}
\value{
A ggplot2 instance that represents a regulation NHL rink
}
\description{
Create a ggplot2 instance that represents a regulation NHL rink, with the
center of the rink corresponding to (0, 0)
}
