% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer-nwsl-features.R
\name{geom_nwsl}
\alias{geom_nwsl}
\title{Create a ggplot2 instance that represents a regulation NWSL pitch, with the
center of the pitch corresponding to (0, 0)}
\usage{
geom_nwsl(
  touchline_length = 120,
  goal_line_length = 90,
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw",
  caption_color = "#707372",
  background_color = NULL,
  ...
)
}
\arguments{
\item{touchline_length}{The entire length of the touchline (both halves).
This can be in any units, but will default to meters. Default: \code{120}}

\item{goal_line_length}{The length of the goal line. This can be in any
units, but will default to meters. Default: \code{90}}

\item{full_surf}{A boolean indicating whether or not to draw a full-surface
representation of the playing surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not the surface representation
needs to be rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the surface
representation. Default: \code{'ccw'}}

\item{caption_color}{A hexadecimal string representing the color to use for
the plot's caption. Default: '#707372' (grey)}

\item{background_color}{A hexadecimal string representing the color to use
for the plot's background. Default: \code{NULL}}

\item{...}{Additional arguments to pass to the function. These should be the
colors to pass to the \code{nwsl_features_set_colors()} function}
}
\value{
A ggplot2 instance that represents a regulation NWSL pitch
}
\description{
Create a ggplot2 instance that represents a regulation NWSL pitch, with the
center of the pitch corresponding to (0, 0)
}
