% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer.R
\name{geom_soccer}
\alias{geom_soccer}
\title{Generate a ggplot2 instance containing a regulation soccer pitch for a
specified league. Since soccer pitches can vary in dimension, the dimensions
must be supplied. These values will default to the maximum values according
to the Laws of the Game}
\usage{
geom_soccer(
  league,
  touchline_length = 120,
  goal_line_length = 90,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{league}{The league for which to draw the surface}

\item{touchline_length}{The length of the touchline. This should be the
entire length (both halves) of the field}

\item{goal_line_length}{The length of the goal line}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: FALSE}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: 'ccw'}
}
\value{
A ggplot2 instance with a full-surface representation of a soccer
pitch
}
\description{
Generate a ggplot2 instance containing a regulation soccer pitch for a
specified league. Since soccer pitches can vary in dimension, the dimensions
must be supplied. These values will default to the maximum values according
to the Laws of the Game
}
\examples{
geom_soccer(league = "MLS")
geom_soccer(league = "PREMIER", rotate = TRUE, rotation_dir = "ccw")
}
