% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport-swimming.R
\name{geom_swimming}
\alias{geom_swimming}
\title{Generate a ggplot2 instance containing a swimming for a specified course}
\usage{
geom_swimming(
  course,
  lane_width = 3,
  number_of_lanes = 8,
  overflow_channels = 1.5,
  rotate = FALSE,
  rotation_dir = "ccw",
  ...
)
}
\arguments{
\item{course}{The course for which to draw the surface}

\item{lane_width}{The width of an individual lane}

\item{number_of_lanes}{The number of lanes in the pool}

\item{overflow_channels}{Width of overflow channels (if they exist)}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}

\item{...}{Additional arguments to pass to the function. These should be the
colors to pass to the \code{{league}_features_set_colors()} function, or
units with which to draw the plot}
}
\value{
A ggplot2 instance with a full-surface representation of a swimming pool
}
\description{
Generate a ggplot2 instance containing a swimming for a specified course
}
\examples{
geom_swimming(course = "SCM")
geom_swimming(league = "LCM", rotate = TRUE, rotation_dir = "ccw")
}
