% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport-tennis.R
\name{geom_tennis}
\alias{geom_tennis}
\title{Generate a ggplot2 instance containing a regulation tennis court for a
specified league}
\usage{
geom_tennis(
  league,
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw",
  ...
)
}
\arguments{
\item{league}{The league for which to draw the surface}

\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}

\item{...}{Additional arguments to pass to the function. These should be the
colors to pass to the \code{{league}_features_set_colors()} function, or
units with which to draw the plot}
}
\description{
Generate a ggplot2 instance containing a regulation tennis court for a
specified league
}
\examples{
geom_tennis(league = "itf")
}
