% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey-features.R
\name{hockey_boards}
\alias{hockey_boards}
\title{The boards are the wall around the outside of the rink that constrain the
playing surface. The boards are typically ovular in shape}
\usage{
hockey_boards(
  rink_length = 0,
  rink_width = 0,
  feature_radius = 0,
  feature_thickness = 0
)
}
\arguments{
\item{rink_length}{The length of the rink}

\item{rink_width}{The width of the rink}

\item{feature_radius}{The radius of the corners of the boards}

\item{feature_thickness}{The thickness with which to draw the boards}
}
\value{
A data frame of the bounding coordinates of the boards
}
\description{
The boards are the wall around the outside of the rink that constrain the
playing surface. The boards are typically ovular in shape
}
\keyword{internal}
