% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-hockey.R
\name{hockey_center_faceoff_circle}
\alias{hockey_center_faceoff_circle}
\title{The center faceoff circle is where the each period of the game begins. It
differs from the non-centered faceoff circles in that there are no adjoining
hash marks on this circle. It is also a different color than the non-centered
faceoff circles. Its line thickness should be given by 'minor_line_thickness'
as this is a minor line on the ice surface}
\usage{
hockey_center_faceoff_circle(feature_radius = 0, feature_thickness = 0)
}
\arguments{
\item{feature_radius}{The radius of the center faceoff circle}

\item{feature_thickness}{The thickness of the line of the center faceoff
circle}
}
\value{
A data frame containing the bounding coordinates of the center
faceoff circle
}
\description{
This draws the line defining the faceoff circle at center ice. The line is
circular in shape, and usually dark blue in color
}
\keyword{internal}
