% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-hockey.R
\name{hockey_center_faceoff_spot}
\alias{hockey_center_faceoff_spot}
\title{The center faceoff spot is the spot at which the game begins. Its center
should lie directly in the center of the ice surface. Its radius is passed as
a key in \code{rink_params}}
\usage{
hockey_center_faceoff_spot(feature_radius = 0)
}
\arguments{
\item{feature_radius}{The radius of the center faceoff spot}
}
\value{
A data frame containing the bounding coordinates of the center
faceoff spot
}
\description{
The center faceoff spot is the spot at which the game begins. Its center
should lie directly in the center of the ice surface. Its radius is passed as
a key in \code{rink_params}
}
\keyword{internal}
