% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey-features.R
\name{hockey_center_line}
\alias{hockey_center_line}
\title{The center line is the line that divides the ice surface in half. Its center
should lie directly in the center of the ice surface. Its line thickness
should be given by 'major_line_thickness' as this is a major line on the ice
surface}
\usage{
hockey_center_line(
  feature_thickness = 0,
  rink_width = 0,
  center_faceoff_spot_gap = 0
)
}
\arguments{
\item{feature_thickness}{The thickness of the center line}

\item{rink_width}{The width of the rink}
}
\value{
A data frame of the bounding coordinates of the center line
}
\description{
The center line is the line that divides the ice surface in half. Its center
should lie directly in the center of the ice surface. Its line thickness
should be given by 'major_line_thickness' as this is a major line on the ice
surface
}
\keyword{internal}
