% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport-hockey.R
\name{hockey_features_set_colors}
\alias{hockey_features_set_colors}
\title{Set the colors to be used for the plot. The values provided in the arguments
are the defaults, and, where specified, are the rule-book specified values.}
\usage{
hockey_features_set_colors(
  plot_background = NULL,
  boards = "#000000",
  ozone_ice = "#ffffff",
  nzone_ice = "#ffffff",
  dzone_ice = "#ffffff",
  center_line = "#c8102e",
  zone_line = "#0033a0",
  goal_line = "#c8102e",
  restricted_trapezoid = "#c8102e",
  goal_crease_outline = "#c8102e",
  goal_crease_fill = "#41b6e6",
  referee_crease = "#c8102e",
  center_faceoff_spot = "#0033a0",
  faceoff_spot_ring = "#c8102e",
  faceoff_spot_stripe = "#c8102e",
  center_faceoff_circle = "#0033a0",
  odzone_faceoff_circle = "#c8102e",
  faceoff_line = "#c8102e",
  goal_frame = "#c8102e",
  goal_fill = "#a5acaf4d",
  team_a_bench = "#ffffff",
  team_b_bench = "#ffffff",
  team_a_penalty_box = "#ffffff",
  team_b_penalty_box = "#ffffff",
  off_ice_officials_box = "#a5acaf"
)
}
\arguments{
\item{plot_background}{A hexadecimal string representing the color to use for
this feature}

\item{boards_color}{A hexadecimal string representing the color to use for
this feature}

\item{center_line_color}{A hexadecimal string representing the color to use
for this feature}

\item{blue_line_color}{A hexadecimal string representing the color to use for
this feature}

\item{goal_line_color}{A hexadecimal string representing the color to use for
this feature}

\item{goalkeepers_restricted_area_color}{A hexadecimal string representing
the color to use for this feature}

\item{goal_crease_outline_color}{A hexadecimal string representing the color
to use for this feature}

\item{goal_crease_fill_color}{A hexadecimal string representing the color to
use for this feature}

\item{referee_crease_color}{A hexadecimal string representing the color to
use for this feature}

\item{center_faceoff_spot_color}{A hexadecimal string representing the color
to use for this feature}

\item{faceoff_spot_outer_ring_color}{A hexadecimal string representing the
color to use for this feature}

\item{faceoff_spot_fill_color}{A hexadecimal string representing the color to
use for this feature}

\item{center_faceoff_circle_color}{A hexadecimal string representing the
color to use for this feature}

\item{non_center_faceoff_circle_color}{A hexadecimal string representing the
color to use for this feature}

\item{faceoff_line_color}{A hexadecimal string representing the color to use
for this feature}

\item{goal_color}{A hexadecimal string representing the color to use for this
feature}

\item{goal_fill_color}{A hexadecimal string representing the color to use for
this feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Hexadecimal values are the passed vales to this function by default, but it
is also possible to use string-named values (e.g. \code{"dodgerblue"}) when
specifying.
}
\keyword{internal}
