% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-hockey.R
\name{hockey_neutral_zone}
\alias{hockey_neutral_zone}
\title{The neutral zone is the middle "third" of the rink. This is the area between
the two zone (blue) lines. The center of the neutral zone should lie along
the line x = 0}
\usage{
hockey_neutral_zone(rink_width = 0, feature_thickness = 0)
}
\arguments{
\item{rink_width}{The width of the rink}

\item{feature_thickness}{The length of the neutral zone}
}
\value{
A data frame containing the bounding coordinates of the neutral zone
}
\description{
The neutral zone is the middle "third" of the rink. This is the area between
the two zone (blue) lines. The center of the neutral zone should lie along
the line x = 0
}
\keyword{internal}
