% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey-features.R
\name{hockey_off_ice_officials_box}
\alias{hockey_off_ice_officials_box}
\title{The off-ice officials' box is located between the two penalty boxes, opposite
the team bench areas}
\usage{
hockey_off_ice_officials_box(
  feature_thickness = 0,
  officials_box_length = 0,
  officials_box_depth = 0
)
}
\arguments{
\item{feature_thickness}{The thickness of the outline of the off-ice
officials' box}

\item{officials_box_length}{The length of the off-ice officials' box}

\item{officials_box_depth}{The depth at which the off-ice officials' box
extends from the outer edge of the boards}
}
\value{
A data frame containing the bounding coordinates of the off-ice
officials' box's outline
}
\description{
This will have the same thickness as the boards, but will be located outside
the ice surface
}
\keyword{internal}
