% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey-features.R
\name{hockey_offensive_zone}
\alias{hockey_offensive_zone}
\title{The offensive zone is the right "third" of the rink in TV view. This is the
area that a team attacks to try to score a goal when attacking from left to
right}
\usage{
hockey_offensive_zone(
  rink_length = 0,
  rink_width = 0,
  feature_radius = 0,
  nzone_length = 0
)
}
\arguments{
\item{rink_length}{The length of the rink}

\item{rink_width}{The width of the rink}

\item{feature_radius}{The radius of the corners of the boards}

\item{nzone_length}{The length of the neutral zone}
}
\value{
A data frame of the bounding coordinates of the offensive zone
}
\description{
The offensive zone is the right "third" of the rink in TV view. This is the
area that a team attacks to try to score a goal when attacking from left to
right
}
\keyword{internal}
