% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey-features.R
\name{hockey_penalty_box_fill}
\alias{hockey_penalty_box_fill}
\title{The penalty boxes are the areas outside the confines of the rink where
players serve time for a penalty incurred. They are to be on the same side of
the ice surface and separate, as close to center ice as possible, for each
team. This will not include the off-ice officials' box; see
\code{\link[=hockey_off_ice_officials_box]{hockey_off_ice_officials_box()}} for more information}
\usage{
hockey_penalty_box_fill(
  feature_thickness = 0,
  penalty_box_length = 0,
  penalty_box_depth = 0
)
}
\arguments{
\item{feature_thickness}{The thickness of the outline of the penalty box}

\item{penalty_box_length}{The length of the penalty box}

\item{penalty_box_depth}{The depth at which the penalty box extends from the
outer edge of the boards}
}
\value{
A data frame containing the bounding coordinates of the penalty box's
inner filling
}
\description{
This will have the same thickness as the boards, but will be located outside
the ice surface
}
\keyword{internal}
