% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-hockey.R
\name{hockey_referee_crease}
\alias{hockey_referee_crease}
\title{The referee's crease is a semi-circle on the "bottom" of the boards (in TV
view), centered on the line y = 0 (the center of the center line)}
\usage{
hockey_referee_crease(feature_radius = 0, feature_thickness = 0)
}
\arguments{
\item{feature_radius}{The radius of the referee's crease}

\item{feature_thickness}{The thickness with which to draw the referee's
crease}
}
\value{
A data frame of the bounding coordinates of the referee's crease
}
\description{
The referee's crease is a semi-circle on the "bottom" of the boards (in TV
view), centered on the line y = 0 (the center of the center line)
}
\keyword{internal}
