% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey-features.R
\name{hockey_zone_line}
\alias{hockey_zone_line}
\title{The zone lines are the lines that separate the neutral zone from the
offensive and defensive zones. Its line thickness should be given by
'major_line_thickness' as this is a major line on the ice surface}
\usage{
hockey_zone_line(rink_width = 0, feature_thickness = 0)
}
\arguments{
\item{rink_width}{The width of the rink}

\item{feature_thickness}{The thickness of the zone line}
}
\value{
A data frame containing the bounding coordinates of the zone line
}
\description{
The zone lines are the lines that separate the neutral zone from the
offensive and defensive zones. Its line thickness should be given by
'major_line_thickness' as this is a major line on the ice surface
}
\keyword{internal}
