% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey-iihf-features.R
\name{iihf_feature_center_line}
\alias{iihf_feature_center_line}
\title{Generate the data frame for the points that comprise the center line}
\usage{
iihf_feature_center_line(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not this feature is needed
for a full-surface representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the center line
}
\description{
Generate the data frame for the points that comprise the center line
}
