% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey-iihf-features.R
\name{iihf_features_set_colors}
\alias{iihf_features_set_colors}
\title{Generate the list of colors for an IIHF rink plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_iihf()} function (or its wrapper \code{geom_hockey()})}
\usage{
iihf_features_set_colors(
  boards_color = "#000000",
  center_line_color = "#c8102e",
  blue_line_color = "#0033a0",
  goal_line_color = "#c8102e",
  goal_crease_outline_color = "#c8102e",
  goal_crease_fill_color = "#41b6e6",
  referee_crease_color = "#c8102e",
  center_faceoff_spot_color = "#0033a0",
  faceoff_spot_outer_ring_color = "#c8102e",
  faceoff_spot_fill_color = "#c8102e",
  center_faceoff_circle_color = "#0033a0",
  non_center_faceoff_circle_color = "#c8102e",
  faceoff_line_color = "#c8102e",
  goal_color = "#c8102e",
  goal_fill_color = "#a5acaf"
)
}
\arguments{
\item{boards_color}{A hexadecimal string representing the color to use for
this feature}

\item{center_line_color}{A hexadecimal string representing the color to use
for this feature}

\item{blue_line_color}{A hexadecimal string representing the color to use for
this feature}

\item{goal_line_color}{A hexadecimal string representing the color to use for
this feature}

\item{goal_crease_outline_color}{A hexadecimal string representing the color to
use for this feature}

\item{goal_crease_fill_color}{A hexadecimal string representing the color to
use for this feature}

\item{referee_crease_color}{A hexadecimal string representing the color to
use for this feature}

\item{center_faceoff_spot_color}{A hexadecimal string representing the color
to use for this feature}

\item{faceoff_spot_outer_ring_color}{A hexadecimal string representing the
color to use for this feature}

\item{faceoff_spot_fill_color}{A hexadecimal string representing the color to
use for this feature}

\item{center_faceoff_circle_color}{A hexadecimal string representing the
color to use for this feature}

\item{non_center_faceoff_circle_color}{A hexadecimal string representing the
color to use for this feature}

\item{faceoff_line_color}{A hexadecimal string representing the color to use
for this feature}

\item{goal_color}{A hexadecimal string representing the color to use for this
feature}

\item{goal_fill_color}{A hexadecimal string representing the color to use for
this feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Generate the list of colors for an IIHF rink plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_iihf()} function (or its wrapper \code{geom_hockey()})
}
