% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.R
\name{in_to_m}
\alias{in_to_m}
\title{Converts inches to meters}
\usage{
in_to_m(inches)
}
\arguments{
\item{inches}{A number of inches to be converted to meters}
}
\value{
The converted number of inches to meters
}
\description{
Converts inches to meters
}
