% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.R
\name{inches_to_yd}
\alias{inches_to_yd}
\title{Converts inches to yards}
\usage{
inches_to_yd(inches)
}
\arguments{
\item{inches}{A number of inches to be converted to yards}
}
\value{
The converted number of inches to yards
}
\description{
Converts inches to yards
}
