% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_hex}
\alias{is_hex}
\title{Identify if a passed color is a hexadecimal string}
\usage{
is_hex(col_str = "")
}
\arguments{
\item{col_str}{The potentially-hexadecimal string to check}
}
\value{
A boolean indicating if the string is in fact hexadecimal
}
\description{
Identify if a passed color is a hexadecimal string
}
\keyword{internal}
