% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennis-itf-features.R
\name{itf_feature_baseline}
\alias{itf_feature_baseline}
\title{Generate the data frame for the points that comprise the baseline(s)}
\usage{
itf_feature_baseline(full_surf = TRUE, rotate = FALSE, rotation_dir = "ccw")
}
\arguments{
\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the baseline(s)
}
\description{
Generate the data frame for the points that comprise the baseline(s)
}
