% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennis-itf-features.R
\name{itf_features_set_colors}
\alias{itf_features_set_colors}
\title{Generate the list of colors for an ITF tennis court plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_itf()} function (or its wrapper \code{geom_tennis()})}
\usage{
itf_features_set_colors(
  court_background_color = "#196f0c",
  baseline_color = "#ffffff",
  doubles_sideline_1_color = "#ffffff",
  doubles_sideline_2_color = "#ffffff",
  singles_sideline_1_color = "#ffffff",
  singles_sideline_2_color = "#ffffff",
  service_line_color = "#ffffff",
  center_service_line_color = "#ffffff",
  center_mark_color = "#ffffff",
  net_color = "#ffffff"
)
}
\arguments{
\item{court_background_color}{A hexadecimal string representing the color to use for
this feature}

\item{baseline_color}{A hexadecimal string representing the color to use for
this feature}

\item{doubles_sideline_1_color}{A hexadecimal string representing the color
to use for this feature}

\item{doubles_sideline_2_color}{A hexadecimal string representing the color
to use for this feature}

\item{singles_sideline_1_color}{A hexadecimal string representing the color
to use for this feature}

\item{singles_sideline_2_color}{A hexadecimal string representing the color
to use for this feature}

\item{service_line_color}{A hexadecimal string representing the color to use
for this feature}

\item{center_service_line_color}{A hexadecimal string representing the color
to use for this feature}

\item{center_mark_color}{A hexadecimal string representing the color to use
for this feature}

\item{net_color}{A hexadecimal string representing the color to use for this
feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Generate the list of colors for an ITF tennis court plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_itf()} function (or its wrapper \code{geom_tennis()})
}
