% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_below_goal_marking}
\alias{lacrosse_below_goal_marking}
\title{On some fields, there is a below goal marking. These are circles that should
mirror each other, with their center points used as anchors}
\usage{
lacrosse_below_goal_marking(below_goal_marking_radius = 0)
}
\arguments{
\item{below_goal_marking_radius}{The radius of the below goal marking}
}
\value{
A data frame containing the bounding coordinates of the below goal
marking
}
\description{
On some fields, there is a below goal marking. These are circles that should
mirror each other, with their center points used as anchors
}
\keyword{internal}
