% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_boards}
\alias{lacrosse_boards}
\title{The boards are the wall around the outside of the field that constrain the
playing surface. The boards are either typically ovular in shape, or not
present if the field is outdoors}
\usage{
lacrosse_boards(
  field_length = 0,
  field_width = 0,
  corner_radius = 0,
  boundary_thickness = 0
)
}
\arguments{
\item{field_length}{The length of the field}

\item{field_width}{The width of the field}

\item{corner_radius}{The radius of the corners of the boards}

\item{boundary_thickness}{The thickness with which to draw the boards}
}
\value{
A data frame of the bounding coordinates of the boards
}
\description{
The boards are the wall around the outside of the field that constrain the
playing surface. The boards are either typically ovular in shape, or not
present if the field is outdoors
}
\keyword{internal}
