% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_center_circle}
\alias{lacrosse_center_circle}
\title{Lacrosse Center Circle}
\usage{
lacrosse_center_circle(center_circle_radius = 0, center_circle_thickness = 0)
}
\arguments{
\item{center_circle_radius}{The outer radius of the center circle}

\item{center_circle_thickness}{The thickness of the center circle}
}
\value{
A data frame containing the bounding coordinates of the center circle
}
\description{
The center circle is where play begins to start a game. This is located at
the center of the field (when present)
}
\keyword{internal}
