% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_change_area_fill}
\alias{lacrosse_change_area_fill}
\title{Lacrosse Change Area (Interior)}
\usage{
lacrosse_change_area_fill(change_area_length = 0, change_area_width = 0)
}
\arguments{
\item{change_area_length}{The length of the change area's interior}

\item{change_area_width}{The distance off the boards that the change area
extends into the playing surface}
}
\value{
A data frame containing the bounding coordinates of the change area's
interior fill
}
\description{
The change area is the box-shaped area in front of the team benches where a
substitutions occur. This feature describes its interior fill; its outline
is controlled by \code{\link[=lacrosse_change_area_outline]{lacrosse_change_area_outline()}}
}
\keyword{internal}
