% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_change_area_outline}
\alias{lacrosse_change_area_outline}
\title{The change area is the box-shaped area in front of the team benches where a
substitutions occur. This feature describes its outline; its interior fill
is controlled by \code{\link[=lacrosse_change_area_fill]{lacrosse_change_area_fill()}}}
\usage{
lacrosse_change_area_outline(
  change_area_length = 0,
  change_area_width = 0,
  feature_thickness = 0
)
}
\arguments{
\item{change_area_length}{The length of the change area's interior}

\item{change_area_width}{The distance off the boards that the change area
extends into the playing surface}

\item{feature_thickness}{The thickness of the outline of the box}
}
\value{
A data frame containing the bounding coordinates of the change area's
outline
}
\description{
The change area is the box-shaped area in front of the team benches where a
substitutions occur. This feature describes its outline; its interior fill
is controlled by \code{\link[=lacrosse_change_area_fill]{lacrosse_change_area_fill()}}
}
\keyword{internal}
