% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_defensive_area_line}
\alias{lacrosse_defensive_area_line}
\title{Lacrosse Defensive Area Line}
\usage{
lacrosse_defensive_area_line(
  defensive_area_line_length = 0,
  line_thickness = 0
)
}
\arguments{
\item{defensive_area_line_length}{The length of the defensive-area line
length}

\item{line_thickness}{The thickness of the defensive area line}
}
\value{
A data frame containing the bounding coordinates of the defensive
area lines
}
\description{
The defensive-area lines run parallel to the sidelines and are connected to
the end line (see \code{\link[=lacrosse_end_line]{lacrosse_end_line()}}) and restraining line (see
\code{\link[=lacrosse_restraining_line]{lacrosse_restraining_line()}}) when these features are present
}
\keyword{internal}
