% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_defensive_zone}
\alias{lacrosse_defensive_zone}
\title{Lacrosse Defensive Zone}
\usage{
lacrosse_defensive_zone(
  field_length = 0,
  field_width = 0,
  corner_radius = 0,
  nzone_length = 0,
  field_shape = "rectangle"
)
}
\arguments{
\item{field_length}{The interior length of the field}

\item{field_width}{The interior width of the field}

\item{corner_radius}{The radius of the corner (assuming \code{field_shape =
"oval"}). Viable options are \code{"rectangle"} or \code{"oval"}}

\item{nzone_length}{The length of the neutral zone}

\item{field_shape}{The shape of the field, passed as a string}
}
\value{
A data frame containing the bounding coordinates of the defensive
zone
}
\description{
The defensive zone is the TV-left area of the playing surface. In many cases,
this will correspond to half of the field's length
}
\keyword{internal}
