% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_end_line}
\alias{lacrosse_end_line}
\title{The end lines run the width of the field, with its interior edge designating
the in-bounds area}
\usage{
lacrosse_end_line(field_width = 0, line_thickness = 0)
}
\arguments{
\item{field_width}{The interior width of the field}

\item{line_thickness}{The thickness of the end line}
}
\value{
A data frame containing the bounding coordinates of the sideline
}
\description{
The end lines run the width of the field, with its interior edge designating
the in-bounds area
}
\keyword{internal}
