% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_face_off_marker}
\alias{lacrosse_face_off_marker}
\title{Lacrosse Faceoff Markers}
\usage{
lacrosse_face_off_marker(
  shape = "O",
  feature_thickness = 0,
  side_length = 0,
  feature_radius = 0
)
}
\arguments{
\item{shape}{one of the following strings (case-insensitive):

\describe{
\item{\code{"X"}}{
An "X"-like shape. This must be specified with the
\code{feature_thickness} parameter to specify the width of each bar of
the "X", and the \code{side_length} parameter to control the length of
each bar
}
\item{\code{"O"}}{
A circle shape. This must be specified with the \code{feature_radius}
parameter to determine the size of the circle
}
}}

\item{feature_thickness}{The thickness of a single bar of the "X" shape}

\item{side_length}{The length of a single bar of the "X" shape}

\item{feature_radius}{The radius of a circular face-off spot}
}
\value{
A data frame containing the bounding coordinates of a face-off spot
}
\description{
The face-off markers are where face-offs occur. They may take one of two
forms: an "X" shape or a circle
}
\keyword{internal}
