% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_field_apron}
\alias{lacrosse_field_apron}
\title{Lacrosse Field Apron}
\usage{
lacrosse_field_apron(
  field_length = 0,
  field_width = 0,
  field_apron_thickness = 0,
  field_shape = "rectangle"
)
}
\arguments{
\item{field_length}{The length of the field}

\item{field_width}{The width of the field}

\item{field_apron_thickness}{The thickness of the field apron}

\item{field_shape}{The shape of the field, passed as a string}
}
\value{
A data frame of the bounding coordinates of the field apron
}
\description{
Outdoor require should have a field apron so that boundary lines are more
visible
}
\keyword{internal}
