% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_goal_arc}
\alias{lacrosse_goal_arc}
\title{Lacrosse Goal Arc}
\usage{
lacrosse_goal_arc(
  goal_arc_extension = 0,
  goal_arc_radius = 0,
  line_thickness = 0
)
}
\arguments{
\item{goal_arc_extension}{The extension from the goal line towards the end
line}

\item{goal_arc_radius}{The outer radius of the goal arc, measured from the
center of the goal line}

\item{line_thickness}{The thickness of the goal arc}
}
\value{
A data frame containing the the bounding coordinates of the goal arc
}
\description{
The arc around the goal circle is a semi-circular area located around the
goal circle (see \code{\link[=lacrosse_goal_circle]{lacrosse_goal_circle()}}) that may extend back to the end
line, but also may be cut off at the goal line extended. The extension should
be controlled via the \code{goal_arc_extension} parameter. Note: the hash
marks are generated via \code{\link[=lacrosse_goal_fan_hash_mark]{lacrosse_goal_fan_hash_mark()}}
}
\keyword{internal}
