% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_goal_arc_hash_mark}
\alias{lacrosse_goal_arc_hash_mark}
\title{The goal arc may have a hash mark that extends towards midfield from the goal
line extended. This is that hash mark, and should be anchored using its outer
edge}
\usage{
lacrosse_goal_arc_hash_mark(goal_arc_hash_mark_length = 0, line_thickness = 0)
}
\arguments{
\item{goal_arc_hash_mark_length}{The length from the goal line extended
towards the center of the field that the hash mark extends}

\item{line_thickness}{The thickness of the hash mark extending from the goal
arc}
}
\value{
A data frame containing the bounding coordinates of the goal arc's
hash mark
}
\description{
The goal arc may have a hash mark that extends towards midfield from the goal
line extended. This is that hash mark, and should be anchored using its outer
edge
}
\keyword{internal}
