% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_goal_circle_fill}
\alias{lacrosse_goal_circle_fill}
\title{Lacrosse Goal Circle (Interior)}
\usage{
lacrosse_goal_circle_fill(
  goal_circle_radius = 0,
  line_thickness = 0,
  goal_circle_full_360 = TRUE,
  goal_depth = 0,
  goal_depth_to_circle = 0
)
}
\arguments{
\item{goal_circle_radius}{The outer radius of the goal circle}

\item{line_thickness}{The thickness of the goal circle}

\item{goal_circle_full_360}{A boolean indicating whether the goal circle
should be a 360 degree circle}

\item{goal_depth}{The depth of the goal}

\item{goal_depth_to_circle}{The distance from the back tip of the goal to the
outer radius of the goal circle}
}
\value{
A data frame containing the bounding coordinates of the goal circle's
enclosed area
}
\description{
This feature is the area enclosed by the goal circle's outline. Please see
\code{\link[=lacrosse_goal_circle]{lacrosse_goal_circle()}} for more information
}
\keyword{internal}
