% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_goal_fan_hash_mark}
\alias{lacrosse_goal_fan_hash_mark}
\title{Lacrosse Goal Fan Hash Mark}
\usage{
lacrosse_goal_fan_hash_mark(
  goal_fan_hash_mark_length = 0,
  line_thickness = 0,
  rotational_angle = 0
)
}
\arguments{
\item{goal_fan_hash_mark_length}{The length of each hash mark along the goal
fan}

\item{line_thickness}{The thickness of each hash mark along the goal fan}

\item{rotational_angle}{The angle (in degrees) that the hash mark should be
rotated}
}
\value{
A data frame containing the bounding coordinates of a hash mark along
the goal fan
}
\description{
The hash marks around the goal fan (see \code{\link[=lacrosse_goal_fan]{lacrosse_goal_fan()}}) are drawn
independently from the goal fan itself. These should just be rectangles with
anchor points along the circle
}
\keyword{internal}
