% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_goal_frame}
\alias{lacrosse_goal_frame}
\title{Lacrosse Goal Frame}
\usage{
lacrosse_goal_frame(
  goal_frame_opening_interior = 0,
  goal_post_thickness = 0,
  goal_depth = 0
)
}
\arguments{
\item{goal_frame_opening_interior}{The interior width of the goal frame's
opening}

\item{goal_post_thickness}{The thickness of the goal post}

\item{goal_depth}{The outer depth at which the goal is anchored}
}
\value{
A data frame containing the bounding coordinates of the goal frame
}
\description{
The goal frame is the posts through which the ball must pass in order to
score a goal. It is usually triangular in shape
}
\keyword{internal}
