% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_goal_mouth}
\alias{lacrosse_goal_mouth}
\title{The goal mouth is similar to the goal fan, except it is plainly a semi-circle
located in front of the goal on fields for which it appears}
\usage{
lacrosse_goal_mouth(goal_mouth_radius = 0, line_thickness = 0)
}
\arguments{
\item{goal_mouth_radius}{The outer radius of the goal mouth}

\item{line_thickness}{The thickness of the line marking the goal mouth}
}
\value{
A data frame containing the bounding coordinates of the goal mouth
}
\description{
The goal mouth is similar to the goal fan, except it is plainly a semi-circle
located in front of the goal on fields for which it appears
}
\keyword{internal}
