% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_goal_net}
\alias{lacrosse_goal_net}
\title{The goal net is the netting that the ball must hit in order to score a point}
\usage{
lacrosse_goal_net(
  goal_frame_opening_interior = 0,
  goal_post_thickness = 0,
  goal_depth = 0
)
}
\arguments{
\item{goal_frame_opening_interior}{The interior width of the goal frame's
opening}

\item{goal_post_thickness}{The thickness of the goal post}

\item{goal_depth}{The outer depth at which the goal is anchored}
}
\value{
A data frame containing the bounding coordinates of the goal net
}
\description{
The goal net is the netting that the ball must hit in order to score a point
}
\keyword{internal}
