% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_neutral_zone}
\alias{lacrosse_neutral_zone}
\title{The neutral zone corresponds to the area between the restraining lines. In
cases where there are no restraining lines, this feature will have 0 length}
\usage{
lacrosse_neutral_zone(nzone_length = 0, field_width = 0)
}
\arguments{
\item{nzone_length}{The length of the neutral zone}

\item{field_width}{The width of the field}
}
\value{
A data frame containing the bounding coordinates of the neutral zone
}
\description{
The neutral zone corresponds to the area between the restraining lines. In
cases where there are no restraining lines, this feature will have 0 length
}
\keyword{internal}
