% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_off_field_officials_box}
\alias{lacrosse_off_field_officials_box}
\title{The off-field officials' box is located between the two penalty boxes,
opposite the team bench areas}
\usage{
lacrosse_off_field_officials_box(
  officials_box_thickness = 0,
  officials_box_length = 0,
  officials_box_depth = 0
)
}
\arguments{
\item{officials_box_thickness}{The thickness of the outline of the off-field
officials' box}

\item{officials_box_length}{The length of the off-field officials' box}

\item{officials_box_depth}{The depth at which the off-field officials' box
extends from the outer edge of the boards}
}
\value{
A data frame containing the bounding coordinates of the off-field
officials' box's outline
}
\description{
This will have the same thickness as the boards, but will be located outside
the field surface
}
\keyword{internal}
