% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_offensive_zone}
\alias{lacrosse_offensive_zone}
\title{Lacrosse Offensive Zone}
\usage{
lacrosse_offensive_zone(
  field_length = 0,
  field_width = 0,
  corner_radius = 0,
  nzone_length = 0,
  field_shape = "rectangle"
)
}
\arguments{
\item{field_length}{The interior length of the field}

\item{field_width}{The interior width of the field}

\item{corner_radius}{The radius of the corner (assuming \code{field_shape =
"oval"}). Viable options are \code{"rectangle"} or \code{"oval"}}

\item{nzone_length}{The length of the neutral zone}

\item{field_shape}{The shape of the field, passed as a string}
}
\value{
A data frame of the bounding coordinates of the offensive zone
}
\description{
The offensive zone is where a team tries to score a goal. It is the TV-right
area on the field
}
\keyword{internal}
