% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_player_bench_outline}
\alias{lacrosse_player_bench_outline}
\title{The player benches are the areas outside the confines of the field where
players not currently on the field are seated. They are to be on the same
side of the field surface and separate, as close to center field as possible}
\usage{
lacrosse_player_bench_outline(
  bench_area_outline_thickness = 0,
  bench_length = 0,
  bench_depth = 0
)
}
\arguments{
\item{bench_area_outline_thickness}{The thickness of the outline of the
player bench areas}

\item{bench_length}{The length of the player bench area}

\item{bench_depth}{The depth of the player bench area}
}
\value{
A data frame containing the bounding coordinates of the player bench
area
}
\description{
This will have the same thickness as the boards, but will be located outside
the field surface. Each bench's outline will share the same color as the
boards
}
\keyword{internal}
