% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_referee_crease}
\alias{lacrosse_referee_crease}
\title{The referee's crease is a semi-circle on the "bottom" of the boards (in TV
view), centered on the line y = 0 (the center of the center line)}
\usage{
lacrosse_referee_crease(referee_crease_radius = 0, line_thickness = 0)

lacrosse_referee_crease(referee_crease_radius = 0, line_thickness = 0)
}
\arguments{
\item{referee_crease_radius}{The radius of the referee's crease}

\item{line_thickness}{The thickness with which to draw the referee's crease}
}
\value{
A data frame of the bounding coordinates of the referee's crease

A data frame of the bounding coordinates of the referee's crease
}
\description{
The referee's crease is a semi-circle on the "bottom" of the boards (in TV
view), centered on the line y = 0 (the center of the center line)

The referee's crease is a semi-circle on the "bottom" of the boards (in TV
view), centered on the line y = 0 (the center of the center line)
}
\keyword{internal}
