% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_restraining_line}
\alias{lacrosse_restraining_line}
\title{Lacrosse Restraining Line}
\usage{
lacrosse_restraining_line(field_width = 0, line_thickness = 0)
}
\arguments{
\item{field_width}{The width of the field}

\item{line_thickness}{The thickness of the restraining line}
}
\value{
A data frame containing the bounding coordinates of the restraining
line
}
\description{
The restraining line spans the entire width of the field (where present) and
connects to the defensive area lines (see \code{\link[=lacrosse_defensive_area_line]{lacrosse_defensive_area_line()}})
}
\keyword{internal}
