% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_sideline}
\alias{lacrosse_sideline}
\title{The sidelines run the length of the field, with its interior edge designating
the in-bounds area}
\usage{
lacrosse_sideline(field_length = 0, line_thickness = 0)
}
\arguments{
\item{field_length}{The interior length of the field}

\item{line_thickness}{The thickness of the sideline}
}
\value{
A data frame containing the bounding coordinates of the sideline
}
\description{
The sidelines run the length of the field, with its interior edge designating
the in-bounds area
}
\keyword{internal}
