% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-lacrosse.R
\name{lacrosse_wing_line}
\alias{lacrosse_wing_line}
\title{The wing lines run parallel to the sidelines and cross the center line}
\usage{
lacrosse_wing_line(wing_line_length = 0, line_thickness = 0)
}
\arguments{
\item{wing_line_length}{The length of the wing line}

\item{line_thickness}{The thickness of the wing line}
}
\value{
A data frame containing the bounding coordinates of the wing line
}
\description{
The wing lines run parallel to the sidelines and cross the center line
}
\keyword{internal}
