% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{league_lookup}
\alias{league_lookup}
\title{Datasets available in this package
A list of lists (JSON) that associates a league code with a given sport that
it can create. This data is primarily useful in the caniplot_league()
function, and is intended to be an internal dataset.}
\format{
An object of class \code{list} of length 19.
}
\usage{
league_lookup
}
\description{
Datasets available in this package
A list of lists (JSON) that associates a league code with a given sport that
it can create. This data is primarily useful in the caniplot_league()
function, and is intended to be an internal dataset.
}
\details{
A list of leagues with their associated sports as a list with a
sublist, or a key-value(s) pairing
}
\keyword{datasets}
