% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-conversions.R
\name{m_to_ft}
\alias{m_to_ft}
\title{Converts meters to feet}
\usage{
m_to_ft(m)
}
\arguments{
\item{m}{A number of meters to be converted to feet}
}
\value{
The converted number of meters to feet
}
\description{
Converts meters to feet
}
