% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-mlb-features.R
\name{mlb_feature_baselines}
\alias{mlb_feature_baselines}
\title{Generate the data frame for the points that comprise the baselines and the
running lane}
\usage{
mlb_feature_baselines()
}
\value{
A list of data frames containing the points that comprise the
baselines and running lane
}
\description{
Generate the data frame for the points that comprise the baselines and the
running lane
}
