% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-mlb-features.R
\name{mlb_feature_bases}
\alias{mlb_feature_bases}
\title{Generate the data frame for the points that comprise the bases}
\usage{
mlb_feature_bases()
}
\value{
A list of data frames containing the points that comprise the bases
}
\description{
Generate the data frame for the points that comprise the bases
}
