% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-mlb-features.R
\name{mlb_feature_batters_boxes}
\alias{mlb_feature_batters_boxes}
\title{Generate the data frame for the points that comprise the batter's boxes and
catcher's box}
\usage{
mlb_feature_batters_boxes()
}
\value{
A list of data frames containing the points that comprise the
batter's boxes and catcher's box added
}
\description{
Generate the data frame for the points that comprise the batter's boxes and
catcher's box
}
