% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-mlb-features.R
\name{mlb_feature_infield_dirt}
\alias{mlb_feature_infield_dirt}
\title{Generate the data frame for the points that comprise the infield dirt and the
dirt circles around home plate}
\usage{
mlb_feature_infield_dirt()
}
\value{
A list of data frames containing the points that comprise the infield
dirt and the dirt circles around home plate
}
\description{
Generate the data frame for the points that comprise the infield dirt and the
dirt circles around home plate
}
