% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-mlb-features.R
\name{mlb_feature_infield_grass}
\alias{mlb_feature_infield_grass}
\title{Generate the data frame for the points that comprise the infield grass}
\usage{
mlb_feature_infield_grass()
}
\value{
A data frame containing the points that comprise the infield grass
}
\description{
Generate the data frame for the points that comprise the infield grass
}
