% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-mlb-features.R
\name{mlb_feature_mound}
\alias{mlb_feature_mound}
\title{Generate the data frame for the points that comprise the pitcher's mound and
pitcher's plate}
\usage{
mlb_feature_mound()
}
\value{
A list of data frames containing the points that comprise the
pitcher's mound and pitcher's plate
}
\description{
Generate the data frame for the points that comprise the pitcher's mound and
pitcher's plate
}
