% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-mlb-features.R
\name{mlb_features_set_colors}
\alias{mlb_features_set_colors}
\title{Generate the list of colors for an MLB field plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_mlb()} function (or its wrapper \code{geom_baseball()})}
\usage{
mlb_features_set_colors(
  infield_dirt_color = "#9b7653",
  home_dirt_color = "#9b7653",
  infield_grass_color = "#395d33",
  mound_color = "#9b7653",
  pitchers_plate_color = "#ffffff",
  home_plate_color = "#ffffff",
  first_base_bag_color = "#ffffff",
  second_base_bag_color = "#ffffff",
  third_base_bag_color = "#ffffff",
  lefty_batters_box_color = "#ffffff",
  righty_batters_box_color = "#ffffff",
  catchers_box_color = "#ffffff",
  rf_line_color = "#ffffff",
  lf_line_color = "#ffffff",
  running_lane_color = "#ffffff"
)
}
\arguments{
\item{infield_dirt_color}{A hexadecimal string representing the color to use
for this feature}

\item{home_dirt_color}{A hexadecimal string representing the color to use for
this feature}

\item{infield_grass_color}{A hexadecimal string representing the color to use
for this feature}

\item{mound_color}{A hexadecimal string representing the color to use for
this feature}

\item{pitchers_plate_color}{A hexadecimal string representing the color to
use for this feature}

\item{home_plate_color}{A hexadecimal string representing the color to use
for this feature}

\item{first_base_bag_color}{A hexadecimal string representing the color to
use for this feature}

\item{second_base_bag_color}{A hexadecimal string representing the color to
use for this feature}

\item{third_base_bag_color}{A hexadecimal string representing the color to
use for this feature}

\item{lefty_batters_box_color}{A hexadecimal string representing the color to
use for this feature}

\item{righty_batters_box_color}{A hexadecimal string representing the color
to use for this feature}

\item{catchers_box_color}{A hexadecimal string representing the color to use
for this feature}

\item{rf_line_color}{A hexadecimal string representing the color to use for
this feature}

\item{lf_line_color}{A hexadecimal string representing the color to use for
this feature}

\item{running_lane_color}{A hexadecimal string representing the color to use
for this feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Generate the list of colors for an MLB field plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_mlb()} function (or its wrapper \code{geom_baseball()})
}
