% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer-mls-features.R
\name{mls_feature_goal_line}
\alias{mls_feature_goal_line}
\title{Generate the data frame for the points that comprise the goal lines}
\usage{
mls_feature_goal_line(
  touchline_length,
  goal_line_length,
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{touchline_length}{The length of the touchline. This should be the
entire length (both halves) of the pitch}

\item{goal_line_length}{The length of the goal line}

\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the goal line(s)
}
\description{
Generate the data frame for the points that comprise the goal lines
}
