% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer-mls-features.R
\name{mls_features_set_colors}
\alias{mls_features_set_colors}
\title{Generate the list of colors for an MLS pitch plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_mls()} function (or its wrapper \code{geom_soccer()})}
\usage{
mls_features_set_colors(
  grass_color = "#196f0c",
  touchline_1_color = "#ffffff",
  touchline_2_color = "#ffffff",
  goal_line_color = "#ffffff",
  halfway_line_color = "#ffffff",
  center_circle_color = "#ffffff",
  corner_circle_1_color = "#ffffff",
  corner_circle_2_color = "#ffffff",
  corner_circle_3_color = "#ffffff",
  corner_circle_4_color = "#ffffff",
  box_5.5m_color = "#ffffff",
  box_9.15m_color = "#ffffff",
  penalty_arc_color = "#ffffff",
  penalty_mark_color = "#ffffff",
  center_mark_color = "#ffffff",
  goal_color = "#ffffff"
)
}
\arguments{
\item{grass_color}{A hexadecimal string representing the color to use for
this feature}

\item{touchline_1_color}{A hexadecimal string representing the color to use
for this feature}

\item{touchline_2_color}{A hexadecimal string representing the color to use
for this feature}

\item{goal_line_color}{A hexadecimal string representing the color to use for
this feature}

\item{halfway_line_color}{A hexadecimal string representing the color to use
for this feature}

\item{center_circle_color}{A hexadecimal string representing the color to use
for this feature}

\item{corner_circle_1_color}{A hexadecimal string representing the color to
use for this feature}

\item{corner_circle_2_color}{A hexadecimal string representing the color to
use for this feature}

\item{corner_circle_3_color}{A hexadecimal string representing the color to
use for this feature}

\item{corner_circle_4_color}{A hexadecimal string representing the color to
use for this feature}

\item{box_5.5m_color}{A hexadecimal string representing the color to use for
this feature}

\item{box_9.15m_color}{A hexadecimal string representing the color to use for
this feature}

\item{penalty_arc_color}{A hexadecimal string representing the color to use
for this feature}

\item{penalty_mark_color}{A hexadecimal string representing the color to use
for this feature}

\item{center_mark_color}{A hexadecimal string representing the color to use
for this feature}

\item{goal_color}{A hexadecimal string representing the color to use for this
feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Generate the list of colors for an MLS pitch plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_mls()} function (or its wrapper \code{geom_soccer()})
}
