% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-nba-features.R
\name{nba_feature_center_circles}
\alias{nba_feature_center_circles}
\title{Generate the data frame for the points that comprise the center circles}
\usage{
nba_feature_center_circles(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not this feature is needed
for a full-surface representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A list of data frames containing the points that comprise the center
circles
}
\description{
Generate the data frame for the points that comprise the center circles
}
