% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-ncaa-bb-features.R
\name{ncaa_bb_feature_restricted_area_arc}
\alias{ncaa_bb_feature_restricted_area_arc}
\title{Generate the data frame for the points that comprise the restricted area arc
underneath the basket}
\usage{
ncaa_bb_feature_restricted_area_arc(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not this feature is needed
for a full-surface representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the restricted area
arc underneath the basket
}
\description{
Generate the data frame for the points that comprise the restricted area arc
underneath the basket
}
