% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-ncaa-bb-features.R
\name{ncaa_bb_feature_three_point_line}
\alias{ncaa_bb_feature_three_point_line}
\title{Generate the data frame for the points that comprise the three-point line and
two-point range}
\usage{
ncaa_bb_feature_three_point_line(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not this feature is needed
for a full-surface representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A list of data frames containing the points that comprise the
three-point line and two-point range
}
\description{
Generate the data frame for the points that comprise the three-point line and
two-point range
}
