% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-ncaa-bb-features.R
\name{ncaa_bb_features_set_colors}
\alias{ncaa_bb_features_set_colors}
\title{Generate the list of colors for an ncaa_bb court plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_ncaa_bb()} function (or its wrapper \code{geom_basketball()})}
\usage{
ncaa_bb_features_set_colors(
  court_background_color = "#d2ab6f",
  center_circle_color = "#000000",
  division_line_color = "#000000",
  endline_color = "#000000",
  sideline_color = "#000000",
  team_bench_color = "#000000",
  substitution_area_color = "#000000",
  court_apron_color = "#d2ab6f",
  m_three_point_line_color = "#000000",
  w_three_point_line_color = "#000000",
  m_two_point_range_color = "#d2ab6f",
  w_two_point_range_color = "#d2ab6f",
  professional_free_throw_lane_color = "#000000",
  professional_painted_area_color = "#d2ab6f",
  amateur_free_throw_lane_color = "#000000",
  amateur_painted_area_color = "#d2ab6f",
  professional_free_throw_lane_lines_color = "#000000",
  amateur_free_throw_lane_lines_color = "#000000",
  free_throw_semi_circle_line_color = "#000000",
  free_throw_semi_circle_fill_color = "#d2ab6f",
  lower_defensive_box_color = "#000000",
  restricted_area_arc_color = "#000000",
  backboard_color = "#000000",
  basket_ring_color = "#000000",
  net_color = "#ffffff"
)
}
\arguments{
\item{court_background_color}{A hexadecimal string representing the color to
use for this feature}

\item{center_circle_color}{A hexadecimal string representing the color to use for this feature}

\item{division_line_color}{A hexadecimal string representing the color to use
for this feature}

\item{endline_color}{A hexadecimal string representing the color to use for
this feature}

\item{sideline_color}{A hexadecimal string representing the color to use for
this feature}

\item{team_bench_color}{A hexadecimal string representing the color to use
for this feature}

\item{substitution_area_color}{A hexadecimal string representing the color to
use for this feature}

\item{court_apron_color}{A hexadecimal string representing the color to use
for this feature}

\item{m_three_point_line_color}{A hexadecimal string representing the color to
use for this feature}

\item{w_three_point_line_color}{A hexadecimal string representing the color to
use for this feature}

\item{m_two_point_range_color}{A hexadecimal string representing the color to
use for this feature}

\item{w_two_point_range_color}{A hexadecimal string representing the color to
use for this feature}

\item{professional_free_throw_lane_color}{A hexadecimal string representing
the color to use for this feature}

\item{professional_painted_area_color}{A hexadecimal string representing the
color to use for this feature}

\item{amateur_free_throw_lane_color}{A hexadecimal string representing the
color to use for this feature}

\item{amateur_painted_area_color}{A hexadecimal string representing the color
to use for this feature}

\item{professional_free_throw_lane_lines_color}{A hexadecimal string
representing the color to use for this feature}

\item{amateur_free_throw_lane_lines_color}{A hexadecimal string representing
the color to use for this feature}

\item{free_throw_semi_circle_line_color}{A hexadecimal string representing
the color to use for this feature}

\item{free_throw_semi_circle_fill_color}{A hexadecimal string representing
the color to use for this feature}

\item{lower_defensive_box_color}{A hexadecimal string representing the color
to use for this feature}

\item{restricted_area_arc_color}{A hexadecimal string representing the color
to use for this feature}

\item{backboard_color}{A hexadecimal string representing the color to use for
this feature}

\item{basket_ring_color}{A hexadecimal string representing the color to use
for this feature}

\item{net_color}{A hexadecimal string representing the color to use for this
feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Generate the list of colors for an ncaa_bb court plot. The defaults can be
overwritten by supplying the names of the list elements to the
\code{geom_ncaa_bb()} function (or its wrapper \code{geom_basketball()})
}
