% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football-ncaa-football-features.R
\name{ncaa_football_feature_directional_arrows_df_maker}
\alias{ncaa_football_feature_directional_arrows_df_maker}
\title{Generate the data frames for the points that comprise the directional arrows.
This function is fed into the ncaa_football_feature_directional_arrows()
function below and helps to optimize performance}
\usage{
ncaa_football_feature_directional_arrows_df_maker(
  yardage,
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw",
  lower = TRUE
)
}
\arguments{
\item{yardage}{The yardage for which to create the yard markings}

\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}

\item{lower}{A boolean indicating whether or not to return the lower
directional arrows}
}
\value{
A data frame (or list of data frames) that contain the coordinates for
the directional arrows
}
\description{
Generate the data frames for the points that comprise the directional arrows.
This function is fed into the ncaa_football_feature_directional_arrows()
function below and helps to optimize performance
}
