% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football-ncaa-football-features.R
\name{ncaa_football_feature_yard_markings}
\alias{ncaa_football_feature_yard_markings}
\title{Generate the data frame for the points that comprise the yard lines}
\usage{
ncaa_football_feature_yard_markings(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the yard lines
}
\description{
Generate the data frame for the points that comprise the yard lines
}
