% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer-ncaa-soccer-features.R
\name{ncaa_soccer_feature_center_circle}
\alias{ncaa_soccer_feature_center_circle}
\title{Generate the data frame for the points that comprise the center circle}
\usage{
ncaa_soccer_feature_center_circle(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the center circle
}
\description{
Generate the data frame for the points that comprise the center circle
}
