% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimming_NCAA_features.R
\name{NCAA_swimming_feature_lane_markers_cross_start}
\alias{NCAA_swimming_feature_lane_markers_cross_start}
\title{Generate the data frame for the points that comprise the lane marker crosses at the start end}
\usage{
NCAA_swimming_feature_lane_markers_cross_start(
  course = "SCY",
  lane_width = 3,
  number_of_lanes = 8,
  overflow_channels = 1,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{course}{The length of the pool as "SCM", "SCY" or "LCM"}

\item{lane_width}{The width of an individual lane}

\item{number_of_lanes}{The number of lanes in the pool}

\item{overflow_channels}{Width of overflow channels (if they exist) in the same units as \code{course}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the lane marker crosses at the start end
}
\description{
Generate the data frame for the points that comprise the lane marker crosses at the start end
}
