% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimming_NCAA_features.R
\name{NCAA_swimming_features_set_colors}
\alias{NCAA_swimming_features_set_colors}
\title{Generate the list of colors for a pool. The defaults can
be overwritten by supplying the names of the list elements to the
\code{geom_swimming()} function}
\usage{
NCAA_swimming_features_set_colors(
  deck_color = "grey",
  pool_color = "blue",
  m15_start_color = "black",
  m15_turn_color = "black",
  m15_markers_color = "red",
  flags_start_color = "red",
  flags_turn_color = "red",
  flags_string_color = "black",
  lane_markers_color = "black",
  blocks_color = "white",
  lane_lines_color = "white",
  lane_line_ends_color = "red",
  lane_line_strings_color = "black"
)
}
\arguments{
\item{deck_color}{A hexadecimal string representing the color to use for
this feature}

\item{pool_color}{A hexadecimal string representing the color to use for
this feature}

\item{lane_markers_color}{A hexadecimal string representing the color to use for
this feature}

\item{m15_color}{A hexadecimal string representing the color to use for
this feature}

\item{flags_color}{A hexadecimal string representing the color to use for
this feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Generate the list of colors for a pool. The defaults can
be overwritten by supplying the names of the list elements to the
\code{geom_swimming()} function
}
