% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennis-ncaa-tennis-features.R
\name{ncaa_tennis_feature_center_service_line}
\alias{ncaa_tennis_feature_center_service_line}
\title{Generate the data frame for the points that comprise the center service
line(s)}
\usage{
ncaa_tennis_feature_center_service_line(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the center service
line(s)
}
\description{
Generate the data frame for the points that comprise the center service
line(s)
}
