% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimming-nfhs-features.R
\name{nfhs_swimming_feature_lane_markers_cross_turn}
\alias{nfhs_swimming_feature_lane_markers_cross_turn}
\title{Generate the data frame for the points that comprise the lane marker crosses at the turn end}
\usage{
nfhs_swimming_feature_lane_markers_cross_turn(
  course = "SCY",
  lane_width = 7/3,
  number_of_lanes = 8,
  overflow_channels = 0.5,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{course}{The length of the pool as "SCM" or "SCY"}

\item{lane_width}{The width of an individual lane}

\item{number_of_lanes}{The number of lanes in the pool}

\item{overflow_channels}{Width of overflow channels (if they exist) in the same units as \code{course}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the lane marker crosses at the turn end
}
\description{
Generate the data frame for the points that comprise the lane marker crosses at the turn end
}
\author{
Gregory A. Pilgrim
}
