% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football-nfl-features.R
\name{nfl_feature_directional_arrows}
\alias{nfl_feature_directional_arrows}
\title{Generate the data frame for the points that comprise the directional arrows}
\usage{
nfl_feature_directional_arrows(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame that contains the points that comprise the directional
arrows
}
\description{
Generate the data frame for the points that comprise the directional arrows
}
