% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football-nfl-features.R
\name{nfl_feature_yard_markings_df_maker}
\alias{nfl_feature_yard_markings_df_maker}
\title{Generate the data frames for the points that comprise the yard lines. This
function is fed into the nfl_feature_yard_markings() function below and helps to
optimize performance}
\usage{
nfl_feature_yard_markings_df_maker(
  yardage,
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{yardage}{The yardage for which to create the yard markings}

\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame (or list of data frames) that contain the coordinates for
the yard lines
}
\description{
Generate the data frames for the points that comprise the yard lines. This
function is fed into the nfl_feature_yard_markings() function below and helps to
optimize performance
}
