% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey-nhl-features.R
\name{nhl_feature_faceoff_circle}
\alias{nhl_feature_faceoff_circle}
\title{Generate the data frame for the points that comprise the faceoff circles}
\usage{
nhl_feature_faceoff_circle(
  center = c(0, 0),
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{center}{The center coordinates of the faceoff spot}

\item{full_surf}{A boolean indicating whether or not this feature is needed
for a full-surface representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the faceoff circle
}
\description{
Generate the data frame for the points that comprise the faceoff circles
}
