% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey-nwhl-features.R
\name{nwhl_feature_goalkeepers_restricted_area}
\alias{nwhl_feature_goalkeepers_restricted_area}
\title{Generate the data frame for the points that comprise the goalkeeper's
restricted area}
\usage{
nwhl_feature_goalkeepers_restricted_area(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not this feature is needed
for a full-surface representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the goalkeeper's
restricted area
}
\description{
Generate the data frame for the points that comprise the goalkeeper's
restricted area
}
