% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer-nwsl-features.R
\name{nwsl_feature_box_5.5m}
\alias{nwsl_feature_box_5.5m}
\title{Generate the data frame for the points that comprise the 6-yard box}
\usage{
nwsl_feature_box_5.5m(
  touchline_length,
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{touchline_length}{The length of the touchline. This should be the
entire length (both halves) of the pitch}

\item{full_surf}{A boolean indicating whether or not to plot a full surface
representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A data frame containing the points that comprise the 6-yard box
}
\description{
Generate the data frame for the points that comprise the 6-yard box
}
