% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{quadratic_formula}
\alias{quadratic_formula}
\title{Deploy the quadratic formula given inputs a, b, and c such that they satisfy the equation (a * (x ^ 2)) + bx + c = 0}
\usage{
quadratic_formula(a, b, c)
}
\arguments{
\item{a}{The coefficient of the squared term}

\item{b}{The coefficient of the linear term}

\item{c}{The coefficient of the constant term}
}
\value{
The solutions to the equation
}
\description{
Deploy the quadratic formula given inputs a, b, and c such that they satisfy the equation (a * (x ^ 2)) + bx + c = 0
}
