% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate-transformations.R
\name{reflect}
\alias{reflect}
\title{Reflect Coordinates}
\usage{
reflect(df, over_x = FALSE, over_y = TRUE)
}
\arguments{
\item{df}{The data frame to reflect. It must have \code{x} and \code{y}
columns}

\item{over_x}{A boolean indicating whether or not to reflect over the x axis.
Default: FALSE}

\item{over_y}{A boolean indicating whether or not to reflect over the y axis.
Default: TRUE}
}
\value{
The reflected data frame
}
\description{
Perform a mathematical reflection of coordinates over a specified axis
}
\examples{
reflect(data.frame(x = 1, y = 0))

}
\keyword{internal}
