% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer.R
\name{soccer_box_18yd}
\alias{soccer_box_18yd}
\title{Generate the dataframe for the points that comprise the 18-yard boxes}
\usage{
soccer_box_18yd(
  g,
  league,
  touchline_length,
  rotate = FALSE,
  rotation_dir = TRUE
)
}
\arguments{
\item{g}{A ggplot2 instance on which to add the feature}

\item{league}{The league for which to draw the surface}

\item{touchline_length}{The length of the touchline. This should be the
entire length (both halves) of the field}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: FALSE}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: 'ccw'}
}
\value{
A ggplot2 instance with the 18-yard boxes added to it
}
\description{
Generate the dataframe for the points that comprise the 18-yard boxes
}
