% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-soccer.R
\name{soccer_corner_arc}
\alias{soccer_corner_arc}
\title{Soccer Corner Arc}
\usage{
soccer_corner_arc(feature_radius = 0, feature_thickness = 0)
}
\arguments{
\item{feature_radius}{The outer radius of the corner arc}

\item{feature_thickness}{The thickness of the corner arc}
}
\value{
A data frame containing the bounding coordinates of the corner arc
}
\description{
The corner arcs are the quarter-circles located where the touchline meets the
goal line
}
\details{
The line thickness will be uniform for all features on the pitch
}
\keyword{internal}
