% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer-features.R
\name{soccer_goal_box}
\alias{soccer_goal_box}
\title{The goal box is the smaller of the two boxes that extend from the goal line
The goal box is usually 5.5 meters (6 yards) from the goal line, but may be
parameterized via this function}
\usage{
soccer_goal_box(
  feature_thickness = 0,
  box_length = 0,
  goal_width = 0,
  goal_post_to_box_edge = 0
)
}
\arguments{
\item{feature_thickness}{The thickness of the goal box}

\item{box_length}{The length of the goal box (from the goal line)}

\item{goal_width}{The interior width of the goal}

\item{goal_post_to_box_edge}{The distance from the interior of the goal post
to the outer edge of the goal box}
}
\value{
A data frame containing the bounding coordinates of the goal box
}
\description{
The line thickness will be uniform for all features on the pitch
}
\keyword{internal}
