% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer.R
\name{soccer_goal_lines}
\alias{soccer_goal_lines}
\title{Generate the dataframe for the points that comprise the goal lines}
\usage{
soccer_goal_lines(
  g,
  league,
  touchline_length,
  goal_line_length,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{g}{A ggplot2 instance on which to add the feature}

\item{league}{The league for which to draw the surface}

\item{touchline_length}{The length of the touchline. This should be the
entire length (both halves) of the field}

\item{goal_line_length}{The length of the goal line}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: FALSE}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: 'ccw'}
}
\value{
A ggplot2 instance with the goal lines added to it
}
\description{
Generate the dataframe for the points that comprise the goal lines
}
