% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-soccer.R
\name{soccer_half_pitch}
\alias{soccer_half_pitch}
\title{Soccer Half Pitch}
\usage{
soccer_half_pitch(pitch_length = 0, pitch_width = 0)
}
\arguments{
\item{pitch_length}{The length of the pitch}

\item{pitch_width}{The width of the pitch}
}
\value{
A data frame containing the bounding coordinates of the half of the
pitch
}
\description{
Half of the pitch is located on each side of the halfway line (see
\code{\link[=soccer_halfway_line]{soccer_halfway_line()}} for more information)
}
\keyword{internal}
