% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-soccer.R
\name{soccer_halfway_line}
\alias{soccer_halfway_line}
\title{The halfway line, aka the midfield line or center line, runs the width of the
pitch, dividing it into two equal halves. The left half (in TV view) will be
the defensive half, and the right half will be the offensive half}
\usage{
soccer_halfway_line(pitch_width = 0, feature_thickness = 0)
}
\arguments{
\item{pitch_width}{The width of the pitch}

\item{feature_thickness}{The thickness of the goal line}
}
\value{
A data frame containing the bounding coordinates of the halfway line
}
\description{
The line thickness will be uniform for all features on the pitch
}
\keyword{internal}
