% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-soccer.R
\name{soccer_penalty_mark}
\alias{soccer_penalty_mark}
\title{Soccer Penalty Mark}
\usage{
soccer_penalty_mark(feature_radius = 0)
}
\arguments{
\item{feature_radius}{The radius of the penalty mark}
}
\value{
A data frame containing the bounding coordinates of the penalty mark
}
\description{
The penalty mark is the center point for the arc of the penalty box, as well
as where any penalty kick is taken from
}
\keyword{internal}
