% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-soccer.R
\name{soccer_touchline}
\alias{soccer_touchline}
\title{The lines that run the full length of the pitch are called the touchlines. In
some cases, they may also be referred to as the sidelines, as they comprise
the sides of the pitch}
\usage{
soccer_touchline(pitch_length = 0, feature_thickness = 0)
}
\arguments{
\item{pitch_length}{The length of the pitch}

\item{feature_thickness}{The thickness of the touchline}
}
\value{
A data frame containing the bounding coordinates of the touchline
}
\description{
The line thickness will be uniform for all features on the pitch
}
\keyword{internal}
