% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-softball.R
\name{softball_base}
\alias{softball_base}
\title{Softball Base}
\usage{
softball_base(
  base_side_length = 0,
  adjust_x_left = FALSE,
  adjust_x_right = FALSE
)
}
\arguments{
\item{base_side_length}{The length of each side of the base}

\item{adjust_x_left}{Whether or not the base should be adjusted in the -x
direction (e.g. first base)}

\item{adjust_x_right}{Whether or not the base should be adjusted in the +x
direction (e.g. third base)}
}
\value{
A data frame that comprises the boundary of the base
}
\description{
One of the bases on the diamond, or really any base on the field. These are
squares that are rotated 45 degrees
}
\keyword{internal}
