% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-softball.R
\name{softball_features_set_colors}
\alias{softball_features_set_colors}
\title{Softball Feature Colors}
\usage{
softball_features_set_colors(
  plot_background = "#395d33",
  infield_dirt = "#9b7653",
  base = "#ffffff",
  pitchers_plate = "#ffffff",
  batters_box = "#ffffff",
  catchers_box = "#ffffff",
  foul_line = "#ffffff",
  running_lane = "#ffffff",
  pitchers_circle = "#ffffff",
  pitchers_lane = "#ffffff",
  coaches_box = "#ffffff"
)
}
\arguments{
\item{plot_background}{A hexadecimal string representing the color to use for
this feature}

\item{infield_dirt}{A hexadecimal string representing the color to use for
this feature}

\item{base}{A hexadecimal string representing the color to use for this
feature}

\item{pitchers_plate}{A hexadecimal string representing the color to use for
this feature}

\item{batters_box}{A hexadecimal string representing the color to use for
this feature}

\item{catchers_box}{A hexadecimal string representing the color to use for
this feature}

\item{foul_line}{A hexadecimal string representing the color to use for this
feature}

\item{running_lane}{A hexadecimal string representing the color to use for
this feature}

\item{pitchers_circle}{A hexadecimal string representing the color to use for
this feature}

\item{pitchers_lane}{A hexadecimal string representing the color to use for
this feature}

\item{coaches_box}{A hexadecimal string representing the color to use for
this feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Set the colors to be used for the plot. The values provided in the arguments
are the defaults, and, where specified, are the rule-book specified values.
}
\details{
Hexadecimal values are the passed vales to this function by default, but it
is also possible to use string-named values (e.g. \code{"dodgerblue"}) when
specifying.
}
\keyword{internal}
