% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-softball.R
\name{softball_pitchers_circle}
\alias{softball_pitchers_circle}
\title{Softball Pitcher's Circle}
\usage{
softball_pitchers_circle(
  pitchers_circle_radius = 0,
  pitchers_circle_thickness = 0
)
}
\arguments{
\item{pitchers_circle_radius}{The radius of the pitcher's circle, measured
to the outer edge of the line}

\item{pitchers_circle_thickness}{The thickness of the chalk line that
comprises the pitcher's circle}
}
\value{
A data frame of the pitcher's circle's bounding coordinates
}
\description{
The pitcher's circle. This is a circular line with its outer edge a fixed
radius from the center of the front edge of the pitcher's plate. The circle
is drawn with its center at the origin, and is anchored to the front edge of
the pitcher's plate when plotted
}
\keyword{internal}
