% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sport_lookup}
\alias{sport_lookup}
\title{A list of lists (JSON) that associates a sport with its various league codes
that can be plotted from its geom_{sport}() function. This data is primarily
useful in the caniplot_sport() function, and is intended to be an internal
dataset.}
\format{
An object of class \code{list} of length 4.
}
\usage{
sport_lookup
}
\description{
A list of lists (JSON) that associates a sport with its various league codes
that can be plotted from its geom_{sport}() function. This data is primarily
useful in the caniplot_sport() function, and is intended to be an internal
dataset.
}
\details{
A list of sports with their associated leagues as a list with a
sublist, or a key-value(s) pairing
}
\keyword{datasets}
