% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surface_dimensions}
\alias{surface_dimensions}
\title{Datasets available in this package
A list of lists (JSON) that associates a league with the appropriate
dimensions to draw its surface. This data follows the current dimensions
specified in the rule books, which can be found in the rule book directory}
\format{
An object of class \code{list} of length 6.
}
\usage{
surface_dimensions
}
\description{
Datasets available in this package
A list of lists (JSON) that associates a league with the appropriate
dimensions to draw its surface. This data follows the current dimensions
specified in the rule books, which can be found in the rule book directory
}
\details{
A set of dimensions by sport and by league
}
\keyword{datasets}
