% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennis-features.R
\name{tennis_backcourt}
\alias{tennis_backcourt}
\title{The backcourt is the area behind the serviceline on the court, contained
within the singles sidelines}
\usage{
tennis_backcourt(court_length = 0, serviceline_distance = 0, singles_width = 0)
}
\arguments{
\item{court_length}{The length of the court}

\item{serviceline_distance}{The distance from the net to the serviceline}

\item{singles_width}{The width of the singles court}
}
\value{
A data frame containing the bounding coordinates of the backcourt
}
\description{
The backcourt is the area behind the serviceline on the court, contained
within the singles sidelines
}
\keyword{internal}
