% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-tennis.R
\name{tennis_baseline}
\alias{tennis_baseline}
\title{Tennis Baseline}
\usage{
tennis_baseline(court_width = 0, feature_thickness = 0)
}
\arguments{
\item{court_width}{The width of the court}

\item{feature_thickness}{The thickness of the baseline}
}
\value{
A data frame containing the bounding coordinates of the baseline
}
\description{
The baseline is the line behind which a player will serve the ball. It spans
the entire width of the court, and its back edge denotes the furthest
boundary inside of which a ball can land and be considered in play
}
\keyword{internal}
