% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennis-features.R
\name{tennis_center_mark}
\alias{tennis_center_mark}
\title{The center mark identifies the center point of the
\code{\link[=tennis_baseline]{tennis_baseline()}}. The line should extend towards the net}
\usage{
tennis_center_mark(center_mark_length = 0, feature_thickness = 0)
}
\arguments{
\item{center_mark_length}{The length of the center mark as measured from the
back edge of the baseline}

\item{feature_thickness}{The thickness of the center mark}
}
\value{
A data frame containing the bounding coordinates of the center mark
}
\description{
The center mark identifies the center point of the
\code{\link[=tennis_baseline]{tennis_baseline()}}. The line should extend towards the net
}
\keyword{internal}
