% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-tennis.R
\name{tennis_center_serviceline}
\alias{tennis_center_serviceline}
\title{The center serviceline on the court divides the service area into two parts:
the ad court (left) and the deuce court (right)}
\usage{
tennis_center_serviceline(center_serviceline_length = 0, feature_thickness = 0)
}
\arguments{
\item{center_serviceline_length}{The length of the center serviceline from
the net to the back edge of the serviceline (see \code{\link[=tennis_serviceline]{tennis_serviceline()}} for
more information)}

\item{feature_thickness}{The thickness of the center serviceline}
}
\value{
A data frame containing the bounding coordinates of the center
serviceline
}
\description{
This line extends from the net to the back edge of the serviceline, and is
centered on the line \code{x = 0}
}
\keyword{internal}
