% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-tennis.R
\name{tennis_court_apron}
\alias{tennis_court_apron}
\title{Tennis Court Apron}
\usage{
tennis_court_apron(
  court_length = 0,
  court_width = 0,
  backstop_distance = 0,
  sidestop_distance = 0
)
}
\arguments{
\item{court_length}{The length of the court}

\item{court_width}{The width of the court (usually the doubles width of the
court)}

\item{backstop_distance}{The distance from the back edge of the
\code{\link[=tennis_baseline]{tennis_baseline()}} to the back boundary}

\item{sidestop_distance}{The distance from the outer edge of the
\code{\link[=tennis_sideline]{tennis_sideline()}} to the side boundary}
}
\value{
A data frame containing the bounding coordinates of the court apron
}
\description{
The court apron is referred to as the backstop and sidestop. These areas are
entirely outside of the playing court, but legal shots made here are
considered in play
}
\keyword{internal}
