% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennis-features.R
\name{tennis_doubles_alley}
\alias{tennis_doubles_alley}
\title{The doubles alley is the area between the singles and doubles sideline. It
should run the entire length of the court}
\usage{
tennis_doubles_alley(court_length = 0, feature_thickness = 0)
}
\arguments{
\item{court_length}{The length of the court}

\item{feature_thickness}{The thickness of the doubles alley (this is the
distance between the singles and doubles sidelines)}
}
\value{
A data frame containing the bounding coordinates of the doubles alley
}
\description{
The doubles alley is the area between the singles and doubles sideline. It
should run the entire length of the court
}
\keyword{internal}
