% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport-tennis.R
\name{tennis_features_set_colors}
\alias{tennis_features_set_colors}
\title{Set the colors to be used for the plot. The values provided in the arguments
are the defaults, and, where specified, are the rule-book specified values.}
\usage{
tennis_features_set_colors(
  plot_background = NULL,
  baseline = "#ffffff",
  singles_sideline = "#ffffff",
  doubles_sideline = "#ffffff",
  serviceline = "#ffffff",
  center_serviceline = "#ffffff",
  center_mark = "#ffffff",
  ad_court = "#395d33",
  deuce_court = "#395d33",
  backcourt = "#395d33",
  doubles_alley = "#395d33",
  court_apron = "#395d33",
  net = "#d3d3d3"
)
}
\arguments{
\item{plot_background}{A hexadecimal string representing the color to use for
this feature}

\item{baseline}{A hexadecimal string representing the color to use for this
feature}

\item{singles_sideline}{A hexadecimal string representing the color to use
for this feature}

\item{doubles_sideline}{A hexadecimal string representing the color to use
for this feature}

\item{serviceline}{A hexadecimal string representing the color to use for
this feature}

\item{center_serviceline}{A hexadecimal string representing the color to use
for this feature}

\item{center_mark}{A hexadecimal string representing the color to use for
this feature}

\item{ad_court}{A hexadecimal string representing the color to use for this
feature}

\item{deuce_court}{A hexadecimal string representing the color to use for
this feature}

\item{backcourt}{A hexadecimal string representing the color to use for this
feature}

\item{doubles_alley}{A hexadecimal string representing the color to use for
this feature}

\item{court_apron}{A hexadecimal string representing the color to use for
this feature}

\item{net}{A hexadecimal string representing the color to use for this
feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Hexadecimal values are the passed vales to this function by default, but it
is also possible to use string-named values (e.g. \code{"dodgerblue"}) when
specifying.
}
\keyword{internal}
