% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-tennis.R
\name{tennis_net}
\alias{tennis_net}
\title{Tennis Net}
\usage{
tennis_net(feature_thickness = 0, net_length = 0)
}
\arguments{
\item{feature_thickness}{The thickness of the net}

\item{net_length}{The length of the net}
}
\value{
A data frame containing the bounding coordinates of the net
}
\description{
The net divides the court into two halves, and should run through the line
\code{x = 0} when viewing the court in TV view
}
\keyword{internal}
