% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-tennis.R
\name{tennis_serviceline}
\alias{tennis_serviceline}
\title{The serviceline is the line in front of which (nearest the net) a serve must
land, and be on the proper side of the court to be considered legal and in
play}
\usage{
tennis_serviceline(singles_width = 0, feature_thickness = 0)
}
\arguments{
\item{singles_width}{The width of the singles court}

\item{feature_thickness}{The thickness of the serviceline}
}
\value{
A data frame containing the bounding coordinates of the serviceline
}
\description{
This line extends completely between the singles sidelines (but not extend to
the doubles sidelines)
}
\keyword{internal}
