% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennis-features.R
\name{tennis_sideline}
\alias{tennis_sideline}
\title{The sideline runs the entire length of the court, and there may be up to four
sidelines on the court (this may refer to either the singles or doubles
sideline). Lines are considered in play, so the outer edge of the sideline
will lie along \code{x = court_width / 2}}
\usage{
tennis_sideline(court_length = 0, feature_thickness = 0)
}
\arguments{
\item{court_length}{The length of the court}

\item{feature_thickness}{The thickness of the sideline}
}
\value{
A data frame containing the bounding coordinates of the sideline
}
\description{
The sideline runs the entire length of the court, and there may be up to four
sidelines on the court (this may refer to either the singles or doubles
sideline). Lines are considered in play, so the outer edge of the sideline
will lie along \code{x = court_width / 2}
}
\keyword{internal}
