% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball.R
\name{three_point_line}
\alias{three_point_line}
\title{Generate the dataframe for the points that comprise the three-point line}
\usage{
three_point_line(
  g,
  league,
  include_m_line = TRUE,
  include_w_line = TRUE,
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{g}{A ggplot2 instance on which to add the feature}

\item{league}{The league for which to draw the surface}

\item{include_m_line}{A boolean indicating whether or not to plot the men's three-point line (NCAA only)}

\item{include_w_line}{A boolean indicating whether or not to plot the women's three-point line (NCAA only)}

\item{full_surf}{A boolean indicating whether or not this feature is needed for a full-surface representation of the surface. Default: TRUE}

\item{rotate}{A boolean indicating whether or not this feature needs to be rotated Default: FALSE}

\item{rotation_dir}{A string indicating which direction to rotate the feature. Default: 'ccw'}
}
\value{
A ggplot2 instance with the three-point line added to it
}
\description{
Generate the dataframe for the points that comprise the three-point line
}
