% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-softball.r
\name{transform_coords}
\alias{transform_coords}
\title{Helper to rotate and translate coordinates}
\usage{
transform_coords(df, rotation = 0, x_trans = 0, y_trans = 0)
}
\description{
Helper to rotate and translate coordinates
}
\keyword{internal}
