% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate-transformations.R
\name{translate}
\alias{translate}
\title{Perform a mathematical translation of coordinates}
\usage{
translate(df, translate_x = 0, translate_y = 0)
}
\arguments{
\item{df}{The data frame to translate It must have \code{x} and \code{y}
columns}

\item{translate_x}{The number of units (in the input data frame's units) to
translate the points in the +x direction. Default: 0}

\item{translate_y}{The number of units (in the input data frame's units) to
translate the points in the +y direction. Default: 0}
}
\value{
The translated data frame
}
\description{
Perform a mathematical translation of coordinates
}
