% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-volleyball.R
\name{volleyball_center_line}
\alias{volleyball_center_line}
\title{Volleyball Center Line}
\usage{
volleyball_center_line(court_width = 0, line_thickness = 0)
}
\arguments{
\item{court_width}{The width of the court, measured from the exterior edges
of the sidelines}

\item{line_thickness}{The thickness of the center line}
}
\value{
A data frame containing the bounding box of the center line
}
\description{
The center line's axis runs along \code{x = 0} when viewing the court in TV
view, dividing the court into two equal halves
}
\keyword{internal}
