% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-volleyball.R
\name{volleyball_end_line}
\alias{volleyball_end_line}
\title{Volleyball End Line}
\usage{
volleyball_end_line(court_width = 0, line_thickness = 0)
}
\arguments{
\item{court_width}{The width of the court, measured from the exterior edges
of the sidelines}

\item{line_thickness}{The thickness of the end line}
}
\value{
A data frame containing the bounding box of the end line
}
\description{
The lines on the court that run the full width of the court are referred to
as the end lines, with the full width of the line being considered in bounds
}
\keyword{internal}
