% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-volleyball.R
\name{volleyball_front_zone}
\alias{volleyball_front_zone}
\title{Volleyball Front Zone}
\usage{
volleyball_front_zone(attack_line_edge_to_center_line = 0, court_width = 0)
}
\arguments{
\item{attack_line_edge_to_center_line}{The distance from the edge furthest
from the attack line to the center of the line running along \code{x = 0}}

\item{court_width}{The width of the court, measured from the exterior edges
of the sidelines}
}
\value{
A data frame containing the bounding coordinates of the center zone
}
\description{
The front zone is the area between the attack line (see
\code{\link[=volleyball_attack_line]{volleyball_attack_line()}}) and the line running along \code{x = 0}. If
considering the entirety of the volleyball court as being divided into
thirds, this is \emph{half} of the middle third of the court
}
\keyword{internal}
