% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-volleyball.R
\name{volleyball_sideline}
\alias{volleyball_sideline}
\title{Volleyball Sideline}
\usage{
volleyball_sideline(court_length = 0, line_thickness = 0)
}
\arguments{
\item{court_length}{The length of the court, measured from the exterior edges
of the end lines}

\item{line_thickness}{The thickness of the sideline}
}
\value{
A data frame containing the bounding box of the sideline
}
\description{
The lines on the court that run the full length of the court are referred to
as the sidelines, with the full width of the line being considered in bounds
}
\keyword{internal}
