% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-volleyball.R
\name{volleyball_substitution_zone_dash}
\alias{volleyball_substitution_zone_dash}
\title{Volleyball Substitution Zone Dash}
\usage{
volleyball_substitution_zone_dash(dash_length = 0, line_thickness = 0)
}
\arguments{
\item{dash_length}{The length of the dash to draw}

\item{line_thickness}{The thickness of the substitution zone dashes}
}
\value{
A data frame containing the bounding box of a single dash of the
substitution zone
}
\description{
The substitution zone is typically marked by a dashed line extending from the
attack lines (see \code{\link[=volleyball_attack_line]{volleyball_attack_line()}} for more). This creates a single
dash, and the dashes should be added to the plot accordingly
}
\keyword{internal}
