% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-wnba-features.R
\name{wnba_feature_free_throw_dashed_semi_circle}
\alias{wnba_feature_free_throw_dashed_semi_circle}
\title{Generate the data frames for the points that comprise the dashed part of the
free-throw circle}
\usage{
wnba_feature_free_throw_dashed_semi_circle(
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{full_surf}{A boolean indicating whether or not this feature is needed
for a full-surface representation of the surface. Default: \code{TRUE}}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: \code{FALSE}}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: \code{'ccw'}}
}
\value{
A list of data frames containing the points that comprise the dashed
part of the free-throw circle
}
\description{
Generate the data frames for the points that comprise the dashed part of the
free-throw circle
}
