% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball-features.R
\name{baseball_infield_grass}
\alias{baseball_infield_grass}
\title{The dirt that comprises the infield grass. This is the area inside the lines
drawn by the basepaths}
\usage{
baseball_infield_grass(
  home_plate_circle_radius = 0,
  foul_line_to_infield_grass = 0,
  baseline_distance = 0,
  base_anchor_to_infield_grass = 0
)
}
\arguments{
\item{home_plate_circle_radius}{The radius of the circle around home plate}

\item{foul_line_to_infield_grass}{The distance from the outer edge of the
foul line to the inner edge of the infield grass}

\item{baseline_distance}{The distance from the back tip of home plate to the
back corner of either first or third base along the foul line}
}
\value{
A data frame that comprises the entirety of the infield grass
}
\description{
The dirt that comprises the infield grass. This is the area inside the lines
drawn by the basepaths
}
\keyword{internal}
