% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball.R
\name{basketball_endline}
\alias{basketball_endline}
\title{Generate the dataframe for the points that comprise the endline(s) line}
\usage{
basketball_endline(
  g,
  league,
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{g}{A ggplot2 instance on which to add the feature}

\item{league}{The league for which to draw the surface}

\item{full_surf}{A boolean indicating whether or not this feature is needed for a full-surface representation of the surface. Default: TRUE}

\item{rotate}{A boolean indicating whether or not this feature needs to be rotated Default: FALSE}

\item{rotation_dir}{A string indicating which direction to rotate the feature. Default: 'ccw'}
}
\value{
A ggplot2 instance with the endline(s) added to it
}
\description{
Generate the dataframe for the points that comprise the endline(s) line
}
