% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cani-functions.R
\name{cani_plot_league}
\alias{cani_plot_league}
\title{Can I Plot League?}
\usage{
cani_plot_league(league_code)
}
\arguments{
\item{league_code}{The case-insensitive league code to be plotted}
}
\value{
Nothing, but a message is sent to the console
}
\description{
Check to see if a league can be plotted, and alert as to which function(s)
that league will work for
}
\examples{
cani_plot_league("MLB")
}
