% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cani-functions.R
\name{cani_plot_sport}
\alias{cani_plot_sport}
\title{Check to see if a sport can be plotted, and alert the user as to which
leagues are plottable for the sport}
\usage{
cani_plot_sport(sport)
}
\arguments{
\item{sport}{The case-insensitive sport name}
}
\value{
Nothing, but a message is sent to the console for the user
}
\description{
Check to see if a sport can be plotted, and alert the user as to which
leagues are plottable for the sport
}
\examples{
cani_plot_sport("basketball")
}
