% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{create_plot_base}
\alias{create_plot_base}
\title{Create the base for the \code{ggplot2} instance with the correct theme
elements}
\usage{
create_plot_base(plot_background = NULL)
}
\arguments{
\item{plot_background}{A hexadecimal string representing the color to use
for the plot's background. Default: \code{NULL}}
}
\value{
A \code{ggplot2} instance onto which the features will be added
}
\description{
Create the base for the \code{ggplot2} instance with the correct theme
elements
}
