% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-shapes.R
\name{create_rectangle}
\alias{create_rectangle}
\title{Create a set of \code{x} and \code{y} coordinates that form a rectangle}
\usage{
create_rectangle(x_min, x_max, y_min, y_max)
}
\arguments{
\item{x_min}{The minimum value of \code{x}}

\item{x_max}{The maximum value of \code{x}}

\item{y_min}{The minimum value of \code{y}}

\item{y_max}{The maximum value of \code{y}}
}
\value{
A data frame containing the points needed to draw the specified
rectangle
}
\description{
Create a set of \code{x} and \code{y} coordinates that form a rectangle
}
