% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-functions.R
\name{distance_formula}
\alias{distance_formula}
\title{Get the Euclidean distance between two points}
\usage{
distance_formula(point_1_x, point_1_y, point_2_x = 0, point_2_y = 0)
}
\arguments{
\item{point_1_x}{A point's (or vector of points') \code{x} coordinate}

\item{point_1_y}{A point's (or vector of points') \code{y} coordinate}

\item{point_2_x}{A point's (or vector of points') \code{x} coordinate}

\item{point_2_y}{A point's (or vector of points') \code{y} coordinate}
}
\value{
The distance between the two supplied points
}
\description{
Get the Euclidean distance between two points
}
\keyword{internal}
