% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport-baseball.R
\name{geom_baseball}
\alias{geom_baseball}
\title{Generate a ggplot2 instance containing a regulation baseball field for a
specified league}
\usage{
geom_baseball(league, ...)
}
\arguments{
\item{league}{The league for which to draw the surface}

\item{...}{Additional arguments to pass to the function. These should be the
colors to pass to the \code{mlb_features_set_colors()} function, or units
with which to draw the plot}
}
\value{
A ggplot2 instance with a full-surface representation of a baseball
field
}
\description{
Generate a ggplot2 instance containing a regulation baseball field for a
specified league
}
\examples{
geom_baseball(league = "MLB")
}
