% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball.R
\name{geom_basketball}
\alias{geom_basketball}
\title{Generate a ggplot2 instance containing a regulation basketball court for a
specified league}
\usage{
geom_basketball(
  league,
  include_amateur = TRUE,
  include_m_line = TRUE,
  include_w_line = TRUE,
  full_surf = TRUE,
  rotate = FALSE,
  rotation_dir = "ccw"
)
}
\arguments{
\item{league}{The league for which to draw the surface}

\item{include_amateur}{A boolean indicating whether or not to include amateur
free-throw lane markings (some professional (W)NBA courts include amateur
free-throw lane markings)}

\item{include_m_line}{A boolean indicating whether or not to plot the men's
three-point line (NCAA only)}

\item{include_w_line}{A boolean indicating whether or not to plot the women's
three-point line (NCAA only)}

\item{full_surf}{A boolean indicating whether or not to plot a full surface
represenation of the surface. Default: TRUE}

\item{rotate}{A boolean indicating whether or not this feature needs to be
rotated. Default: FALSE}

\item{rotation_dir}{A string indicating which direction to rotate the
feature. Default: 'ccw'}
}
\value{
A ggplot2 instance with a full-surface representation of a basketball
court
}
\description{
Generate a ggplot2 instance containing a regulation basketball court for a
specified league
}
\examples{
geom_basketball(league = "NBA")
geom_basketball(league = "NCAA", full_surf = FALSE)
geom_basketball(league = "FIBA", rotate = TRUE, rotation_dir = "ccw")
}
