% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey.R
\name{geom_hockey}
\alias{geom_hockey}
\title{Generate a ggplot2 instance containing an ice rink for a specified league}
\usage{
geom_hockey(league, full_surf = TRUE, rotate = FALSE, rotation_dir = "ccw")
}
\arguments{
\item{league}{The league for which to draw the surface. Default: NHL}

\item{full_surf}{A boolean indicating whether or not to plot a full surface represenation of the surface. Default: TRUE}

\item{rotate}{A boolean indicating whether or not this feature needs to be rotated
Default: FALSE}

\item{rotation_dir}{A string indicating which direction to rotate the feature. Default: 'ccw'}
}
\value{
A ggplot2 instance with the boards added to it
}
\description{
Generate a ggplot2 instance containing an ice rink for a specified league
}
\examples{
geom_hockey(league = "NHL")
geom_hockey(league = "IIHF", full_surf = FALSE)
geom_hockey(league = "NCAA", rotate = TRUE, rotation_dir = "ccw")
}
