% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer-features.R
\name{soccer_center_mark}
\alias{soccer_center_mark}
\title{The center mark is where kickoffs for each half, as well as following any
goal, are taken. The radius should be given to the outside of the mark. This
feature is located at midfield}
\usage{
soccer_center_mark(feature_radius = 0)
}
\arguments{
\item{feature_radius}{The radius of the center mark on the pitch}
}
\value{
A data frame containing the bounding coordinates of the center mark
}
\description{
The center mark is where kickoffs for each half, as well as following any
goal, are taken. The radius should be given to the outside of the mark. This
feature is located at midfield
}
\keyword{internal}
