% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-soccer.R
\name{soccer_goal_line}
\alias{soccer_goal_line}
\title{The goal line is the line that runs the full width of the pitch. The ball
must completely cross the goal line to score a goal for the attacking team}
\usage{
soccer_goal_line(pitch_width = 0, feature_thickness = 0)
}
\arguments{
\item{pitch_width}{The width of the pitch}

\item{feature_thickness}{The thickness of the goal line}
}
\value{
A data frame containing the bounding coordinates of the goal line
}
\description{
The line thickness will be uniform for all features on the pitch
}
\keyword{internal}
