% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surface_dimensions}
\alias{surface_dimensions}
\title{surface_dimensions}
\description{
A list of lists (JSON) that associates a league with the appropriate
dimensions to draw its surface. This data follows the current dimensions
specified in the rule books, which can be found in the rule book directory
}
\details{
A set of dimensions by sport and by league
}
