% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transformations.R
\name{translate}
\alias{translate}
\title{Perform a mathematical translation of coordinates}
\usage{
translate(df, translate_x = 0, translate_y = 0)
}
\arguments{
\item{df}{The dataframe to translate It must have \code{x} and \code{y} columns}

\item{translate_x}{The number of units (in the input dataframe's units) to translate the points in the +x direction. Default: 0}

\item{translate_y}{The number of units (in the input dataframe's units) to translate the points in the +y direction. Default: 0}
}
\value{
The translated dataframe
}
\description{
Perform a mathematical translation of coordinates
}
