% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-volleyball.R
\name{volleyball_features_set_colors}
\alias{volleyball_features_set_colors}
\title{Set the colors to be used for the plot. The values provided in the arguments
are the defaults, and, where specified, are the rule-book specified values.}
\usage{
volleyball_features_set_colors(
  plot_background = NULL,
  free_zone = "#d2ab6f",
  front_zone = "#d2ab6f",
  defensive_backcourt = "#d2ab6f",
  offensive_backcourt = "#d2ab6f",
  court_apron = "#13294b",
  end_line = "#ffffff",
  sideline = "#ffffff",
  attack_line = "#ffffff",
  center_line = "#ffffff",
  service_zone_mark = "#ffffff",
  substitution_zone = "#ffffff"
)
}
\arguments{
\item{plot_background}{A hexadecimal string representing the color to use for
this feature}

\item{free_zone}{A hexadecimal string representing the color to use for this
feature}

\item{front_zone}{A hexadecimal string representing the color to use for this
feature}

\item{defensive_backcourt}{A hexadecimal string representing the color to use
for this feature}

\item{offensive_backcourt}{A hexadecimal string representing the color to use
for this feature}

\item{court_apron}{A hexadecimal string representing the color to use for
this feature}

\item{end_line}{A hexadecimal string representing the color to use for this
feature}

\item{sideline}{A hexadecimal string representing the color to use for this
feature}

\item{attack_line}{A hexadecimal string representing the color to use for
this feature}

\item{center_line}{A hexadecimal string representing the color to use for
this feature}

\item{service_zone_mark}{A hexadecimal string representing the color to use
for this feature}

\item{substitution_zone}{A hexadecimal string representing the color to use
for this feature}
}
\value{
A list of hexadecimal colors to use to color the features on the
resulting plot
}
\description{
Hexadecimal values are the passed vales to this function by default, but it
is also possible to use string-named values (e.g. \code{"dodgerblue"}) when
specifying.
}
\keyword{internal}
